+++ crates/opennote-desktop/src/globals/velotype.rs
@@ -3,7 +3,6 @@ use gpui::App;
 
 use opennote_core_logics::configurations::{ApplicationType, get_configuration_folder_path};
 use opennote_velotype::{
-    app_menu,
     components::init_with_keybindings,
     config::preferences::{EditorSettings, load_or_create_app_preferences_with_path},
     i18n::I18nManager,
@@ -21,7 +20,6 @@ pub fn init_velotype(cx: &mut App) -> Result<()> {
     EditorSettings::init(cx, preferences.show_table_headers);
     install_http_client(cx);
     init_with_keybindings(cx, &preferences.keybindings);
-    app_menu::init(cx);
 
     Ok(())
 }
