--- crates/opennote-velotype/src/app_identity.rs
@@ -1,4 +1,0 @@
-//! Shared application identity used by platform windowing and packaging.
-
-/// Reverse-DNS application id used by GPUI, desktop launchers, and bundles.
-pub(crate) const VELOTYPE_APP_ID: &str = "com.manyougz.Velotype";
