+++ crates/opennote-velotype/src/components/mod.rs
@@ -5,11 +5,8 @@ mod block;
 pub(crate) mod latex;
 pub(crate) mod markdown;
 pub(crate) mod mermaid;
-pub(crate) mod switch;
 
 pub use crate::editor::Editor;
-#[allow(unused_imports)]
-pub(crate) use crate::editor::InfoDialogKind;
 pub use actions::*;
 pub use block::*;
 #[allow(unused_imports)]
