+++ crates/opennote-velotype/src/config/mod.rs
@@ -8,12 +8,6 @@ use serde_json::{Map, Value};
 
 pub mod preferences;
 
-pub(crate) use preferences::{
-    EditorSettings, ImagePasteBehavior, apply_configured_language, apply_configured_theme,
-    import_language_config_and_select, import_theme_config_and_select, open_preferences_window,
-    read_app_preferences,
-};
-
 pub(crate) const RECENT_FILES_LIMIT: usize = 20;
 
 /// Cross-platform configuration directories owned by Velotype.
@@ -67,14 +61,6 @@ pub(crate) fn read_recent_files() -> anyhow::Result<Vec<PathBuf>> {
     read_recent_files_with_dirs(&VelotypeConfigDirs::from_system()?)
 }
 
-pub(crate) fn record_recent_file(path: &Path) -> anyhow::Result<Vec<PathBuf>> {
-    record_recent_file_with_dirs(path, &VelotypeConfigDirs::from_system()?)
-}
-
-pub(crate) fn remove_recent_file(path: &Path) -> anyhow::Result<Vec<PathBuf>> {
-    remove_recent_file_with_dirs(path, &VelotypeConfigDirs::from_system()?)
-}
-
 pub(crate) fn read_recent_files_with_dirs(
     dirs: &VelotypeConfigDirs,
 ) -> anyhow::Result<Vec<PathBuf>> {
@@ -88,67 +74,6 @@ pub(crate) fn read_recent_files_with_dirs(
     };
 
     Ok(normalize_recent_files(text.lines().map(PathBuf::from)))
-}
-
-pub(crate) fn record_recent_file_with_dirs(
-    path: &Path,
-    dirs: &VelotypeConfigDirs,
-) -> anyhow::Result<Vec<PathBuf>> {
-    if path.to_string_lossy().trim().is_empty() {
-        bail!("recent file path cannot be empty");
-    }
-    if !is_recordable_recent_file_path(path) {
-        return read_recent_files_with_dirs(dirs);
-    }
-
-    let mut paths = read_recent_files_with_dirs(dirs)?;
-    let path = path.to_path_buf();
-    paths.retain(|existing| !same_recent_path(existing, &path));
-    paths.insert(0, path);
-    paths.truncate(RECENT_FILES_LIMIT);
-    write_recent_files_with_dirs(&paths, dirs)?;
-    Ok(paths)
-}
-
-pub(crate) fn remove_recent_file_with_dirs(
-    path: &Path,
-    dirs: &VelotypeConfigDirs,
-) -> anyhow::Result<Vec<PathBuf>> {
-    let mut paths = read_recent_files_with_dirs(dirs)?;
-    paths.retain(|existing| !same_recent_path(existing, path));
-    write_recent_files_with_dirs(&paths, dirs)?;
-    Ok(paths)
-}
-
-fn write_recent_files_with_dirs(
-    paths: &[PathBuf],
-    dirs: &VelotypeConfigDirs,
-) -> anyhow::Result<()> {
-    let history_file = dirs.history_file();
-    let normalized = normalize_recent_files(paths.iter().cloned());
-    if normalized.is_empty() {
-        match std::fs::remove_file(&history_file) {
-            Ok(()) => {}
-            Err(err) if err.kind() == std::io::ErrorKind::NotFound => {}
-            Err(err) => {
-                return Err(err)
-                    .with_context(|| format!("failed to remove '{}'", history_file.display()));
-            }
-        }
-        return Ok(());
-    }
-
-    if let Some(parent) = history_file.parent() {
-        std::fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create '{}'", parent.display()))?;
-    }
-    let mut content = String::new();
-    for path in normalized {
-        content.push_str(&path.to_string_lossy());
-        content.push('\n');
-    }
-    std::fs::write(&history_file, content)
-        .with_context(|| format!("failed to write '{}'", history_file.display()))
 }
 
 fn normalize_recent_files(paths: impl IntoIterator<Item = PathBuf>) -> Vec<PathBuf> {
@@ -402,228 +327,5 @@ fn is_empty_json_value(value: &Value) -> bool {
         Value::Array(items) => items.iter().all(is_empty_json_value),
         Value::Object(object) => object.values().all(is_empty_json_value),
         Value::Bool(_) | Value::Number(_) => false,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        RECENT_FILES_LIMIT, VelotypeConfigDirs, parse_jsonc_value, prune_empty_json_values,
-        read_recent_files_with_dirs, record_recent_file_with_dirs, remove_recent_file_with_dirs,
-        sanitize_config_file_stem, strip_jsonc_comments,
-    };
-    use serde_json::json;
-    use std::path::{Path, PathBuf};
-
-    #[test]
-    fn jsonc_comments_are_stripped_without_touching_strings() {
-        let text = r#"
-        {
-            // line comment
-            "url": "https://example.com/a//b",
-            "text": "/* not a comment */",
-            /* block comment */
-            "value": 1
-        }
-        "#;
-
-        let parsed = parse_jsonc_value(text).expect("jsonc should parse");
-        assert_eq!(parsed["url"], "https://example.com/a//b");
-        assert_eq!(parsed["text"], "/* not a comment */");
-        assert_eq!(parsed["value"], 1);
-        assert!(strip_jsonc_comments(text).is_ok());
-    }
-
-    #[test]
-    fn empty_values_are_pruned_recursively() {
-        let mut value = json!({
-            "name": "",
-            "colors": {
-                "text_default": null,
-                "selection": "#fff"
-            },
-            "items": ["", null]
-        });
-
-        assert!(!prune_empty_json_values(&mut value));
-        assert_eq!(value, json!({ "colors": { "selection": "#fff" } }));
-    }
-
-    #[test]
-    fn config_file_stems_are_sanitized() {
-        assert_eq!(
-            sanitize_config_file_stem("My Theme / Blue"),
-            "My_Theme_Blue"
-        );
-        assert_eq!(sanitize_config_file_stem("  ...  "), "custom");
-    }
-
-    #[test]
-    fn missing_recent_history_file_returns_empty_list() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-
-        assert!(read_recent_files_with_dirs(&dirs).unwrap().is_empty());
-        assert!(!dirs.history_file().exists());
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn empty_recent_history_write_does_not_create_file() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-
-        super::write_recent_files_with_dirs(&[], &dirs).unwrap();
-
-        assert!(!dirs.history_file().exists());
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn blank_recent_file_path_is_rejected() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-
-        assert!(record_recent_file_with_dirs(Path::new("   "), &dirs).is_err());
-        assert!(!dirs.history_file().exists());
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn recent_history_filters_empty_lines_and_deduplicates() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        std::fs::create_dir_all(&root).unwrap();
-        std::fs::write(
-            dirs.history_file(),
-            "  \nC:\\one.md\n\nC:\\two.md\nC:\\one.md\n",
-        )
-        .unwrap();
-
-        let paths = read_recent_files_with_dirs(&dirs).unwrap();
-        assert_eq!(
-            paths,
-            vec![PathBuf::from("C:\\one.md"), PathBuf::from("C:\\two.md")]
-        );
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn recent_history_filters_legacy_velotype_temp_fixture_paths() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let fixture_path = std::env::temp_dir().join(format!(
-            "velotype-drop-save-replace-{}-123.md",
-            std::process::id()
-        ));
-        let real_path = PathBuf::from("C:\\notes\\real.md");
-        std::fs::create_dir_all(&root).unwrap();
-        std::fs::write(
-            dirs.history_file(),
-            format!("{}\n{}\n", fixture_path.display(), real_path.display()),
-        )
-        .unwrap();
-
-        let paths = read_recent_files_with_dirs(&dirs).unwrap();
-        assert_eq!(paths, vec![real_path]);
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn recording_velotype_temp_fixture_path_is_noop() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let fixture_path = std::env::temp_dir().join(format!(
-            "velotype-drop-dirty-discard-{}-123.md",
-            std::process::id()
-        ));
-
-        assert!(
-            record_recent_file_with_dirs(&fixture_path, &dirs)
-                .unwrap()
-                .is_empty()
-        );
-        assert!(!dirs.history_file().exists());
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn ordinary_temp_markdown_file_can_still_be_recorded() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let path = std::env::temp_dir().join(format!("manual-note-{}.md", std::process::id()));
-
-        let paths = record_recent_file_with_dirs(&path, &dirs).unwrap();
-
-        assert_eq!(paths, vec![path]);
-        assert!(dirs.history_file().exists());
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn recording_recent_file_moves_it_to_front_and_truncates() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-
-        for index in 0..(RECENT_FILES_LIMIT + 2) {
-            record_recent_file_with_dirs(&PathBuf::from(format!("file-{index}.md")), &dirs)
-                .unwrap();
-        }
-        record_recent_file_with_dirs(&PathBuf::from("file-3.md"), &dirs).unwrap();
-
-        let paths = read_recent_files_with_dirs(&dirs).unwrap();
-        assert_eq!(paths.len(), RECENT_FILES_LIMIT);
-        assert_eq!(paths[0], PathBuf::from("file-3.md"));
-        assert_eq!(
-            paths
-                .iter()
-                .filter(|path| path.as_path() == Path::new("file-3.md"))
-                .count(),
-            1
-        );
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn removing_recent_file_persists_history_without_it() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        record_recent_file_with_dirs(&PathBuf::from("one.md"), &dirs).unwrap();
-        record_recent_file_with_dirs(&PathBuf::from("two.md"), &dirs).unwrap();
-
-        let paths = remove_recent_file_with_dirs(&PathBuf::from("one.md"), &dirs).unwrap();
-
-        assert_eq!(paths, vec![PathBuf::from("two.md")]);
-        assert_eq!(
-            read_recent_files_with_dirs(&dirs).unwrap(),
-            vec![PathBuf::from("two.md")]
-        );
-
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn removing_last_recent_file_deletes_history_file() {
-        let root = std::env::temp_dir().join(format!("velotype-config-{}", uuid::Uuid::new_v4()));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let path = PathBuf::from("only.md");
-        record_recent_file_with_dirs(&path, &dirs).unwrap();
-        assert!(dirs.history_file().exists());
-
-        let paths = remove_recent_file_with_dirs(&path, &dirs).unwrap();
-
-        assert!(paths.is_empty());
-        assert!(!dirs.history_file().exists());
-        assert!(read_recent_files_with_dirs(&dirs).unwrap().is_empty());
-
-        let _ = std::fs::remove_dir_all(root);
     }
 }
