+++ crates/opennote-velotype/src/config/preferences.rs
@@ -4,21 +4,12 @@ use std::collections::BTreeMap;
 use std::path::PathBuf;
 
 use anyhow::Context as _;
-use gpui::prelude::FluentBuilder;
 use gpui::*;
 use serde::{Deserialize, Serialize};
 
 use super::{VelotypeConfigDirs, read_recent_files};
-use crate::components::{
-    ShortcutCategory, ShortcutCommand, ShortcutDefinition, install_keybindings,
-    normalize_shortcut_config, normalize_shortcut_keys, resolved_shortcut_keys,
-    shortcut_conflict_for, shortcut_definitions, switch::Switch,
-};
-use crate::i18n::{I18nManager, language_id_for_locale_preferences};
-use crate::theme::{Theme, ThemeCatalogEntry, ThemeManager};
-use crate::window_chrome::{
-    custom_titlebar_height, render_custom_titlebar, velotype_window_options,
-};
+use crate::components::normalize_shortcut_config;
+use crate::i18n::language_id_for_locale_preferences;
 
 const DEFAULT_THEME_ID: &str = "velotype";
 const DEFAULT_LANGUAGE_ID: &str = "en-US";
@@ -512,1799 +503,4 @@ pub(crate) fn save_app_preferences_with_dirs(
 pub(crate) fn first_existing_recent_markdown_file() -> Option<PathBuf> {
     let recent_files = read_recent_files().ok()?;
     recent_files.into_iter().find(|path| path.is_file())
-}
-
-pub(crate) fn apply_configured_language(cx: &mut App, language_id: &str) -> anyhow::Result<bool> {
-    let mut applied = false;
-    let changed = cx.update_global::<I18nManager, _>(|i18n_manager, _cx| {
-        let changed = i18n_manager.set_language_by_id(language_id);
-        applied = changed || i18n_manager.current_language_id() == language_id;
-        changed
-    });
-    if !applied {
-        return Ok(false);
-    }
-    update_app_preferences(|preferences| {
-        preferences.default_language_id = language_id.into();
-    })?;
-    Ok(changed)
-}
-
-pub(crate) fn apply_configured_theme(cx: &mut App, theme_id: &str) -> anyhow::Result<bool> {
-    let mut applied = false;
-    let changed = cx.update_global::<ThemeManager, _>(|theme_manager, _cx| {
-        let changed = theme_manager.set_theme_by_id(theme_id);
-        applied = changed || theme_manager.current_theme_id() == theme_id;
-        changed
-    });
-    if !applied {
-        return Ok(false);
-    }
-    update_app_preferences(|preferences| {
-        preferences.default_theme_id = theme_id.into();
-    })?;
-    Ok(changed)
-}
-
-pub(crate) fn import_language_config_and_select(
-    cx: &mut App,
-    path: impl AsRef<std::path::Path>,
-) -> anyhow::Result<String> {
-    let imported_id = cx.update_global::<I18nManager, _>(|i18n_manager, _cx| {
-        i18n_manager.import_language_config(path)
-    })?;
-    update_app_preferences(|preferences| {
-        preferences.default_language_id = imported_id.clone();
-    })?;
-    Ok(imported_id)
-}
-
-pub(crate) fn import_theme_config_and_select(
-    cx: &mut App,
-    path: impl AsRef<std::path::Path>,
-) -> anyhow::Result<String> {
-    let imported_id = cx.update_global::<ThemeManager, _>(|theme_manager, _cx| {
-        theme_manager.import_theme_config(path)
-    })?;
-    update_app_preferences(|preferences| {
-        preferences.default_theme_id = imported_id.clone();
-    })?;
-    Ok(imported_id)
-}
-
-pub(crate) fn save_preferences_from_window(
-    startup_open: StartupOpenPreference,
-    default_theme_id: &str,
-    image_paste_behavior: ImagePasteBehavior,
-    keybindings: BTreeMap<String, Vec<String>>,
-    status_bar: &StatusBarPreferences,
-) -> anyhow::Result<AppPreferences> {
-    let dirs = VelotypeConfigDirs::from_system()?;
-    save_preferences_from_window_with_dirs(
-        startup_open,
-        default_theme_id,
-        image_paste_behavior,
-        keybindings,
-        status_bar,
-        &dirs,
-    )
-}
-
-fn save_preferences_from_window_with_dirs(
-    startup_open: StartupOpenPreference,
-    default_theme_id: &str,
-    image_paste_behavior: ImagePasteBehavior,
-    keybindings: BTreeMap<String, Vec<String>>,
-    status_bar: &StatusBarPreferences,
-    dirs: &VelotypeConfigDirs,
-) -> anyhow::Result<AppPreferences> {
-    let mut preferences =
-        load_or_create_app_preferences_with_dirs_and_locales(dirs, sys_locale::get_locales())?;
-    preferences.startup_open = startup_open;
-    preferences.default_theme_id = default_theme_id.into();
-    preferences.image_paste_behavior = image_paste_behavior;
-    preferences.keybindings = normalize_shortcut_config(&keybindings);
-    preferences.status_bar = status_bar.clone();
-    save_app_preferences_with_dirs(&preferences, dirs)?;
-    Ok(preferences)
-}
-
-fn update_app_preferences(
-    update: impl FnOnce(&mut AppPreferences),
-) -> anyhow::Result<AppPreferences> {
-    let mut preferences = load_or_create_app_preferences()?;
-    update(&mut preferences);
-    save_app_preferences(&preferences)?;
-    Ok(preferences)
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum PreferencesNav {
-    File,
-    Theme,
-    Image,
-    Shortcuts,
-    StatusBar,
-}
-
-/// Independent preferences window view.
-pub(crate) struct PreferencesWindow {
-    nav: PreferencesNav,
-    startup_open: StartupOpenPreference,
-    selected_theme_id: String,
-    image_paste_behavior: ImagePasteBehavior,
-    keybindings: BTreeMap<String, Vec<String>>,
-    saved_startup_open: StartupOpenPreference,
-    saved_theme_id: String,
-    saved_image_paste_behavior: ImagePasteBehavior,
-    saved_keybindings: BTreeMap<String, Vec<String>>,
-    theme_options: Vec<ThemeCatalogEntry>,
-    focus_handle: FocusHandle,
-    startup_dropdown_open: bool,
-    theme_dropdown_open: bool,
-    image_dropdown_open: bool,
-    recording_shortcut: Option<ShortcutCommand>,
-    shortcut_error: Option<String>,
-    status_bar_enabled: bool,
-    status_bar_show_word_count: bool,
-    status_bar_show_cursor_position: bool,
-    status_bar_show_sidebar_toggle: bool,
-    status_bar_show_mode_switch: bool,
-    saved_status_bar_enabled: bool,
-    saved_status_bar_show_word_count: bool,
-    saved_status_bar_show_cursor_position: bool,
-    saved_status_bar_show_sidebar_toggle: bool,
-    saved_status_bar_show_mode_switch: bool,
-}
-
-impl PreferencesWindow {
-    fn new(
-        preferences: AppPreferences,
-        theme_options: Vec<ThemeCatalogEntry>,
-        cx: &mut Context<Self>,
-    ) -> Self {
-        let selected_theme_id = if theme_options
-            .iter()
-            .any(|entry| entry.id == preferences.default_theme_id)
-        {
-            preferences.default_theme_id
-        } else {
-            DEFAULT_THEME_ID.into()
-        };
-        let startup_open = preferences.startup_open;
-        let image_paste_behavior = preferences.image_paste_behavior;
-        let keybindings = preferences.keybindings;
-        Self {
-            nav: PreferencesNav::File,
-            startup_open,
-            selected_theme_id: selected_theme_id.clone(),
-            image_paste_behavior,
-            keybindings: keybindings.clone(),
-            saved_startup_open: startup_open,
-            saved_theme_id: selected_theme_id,
-            saved_image_paste_behavior: image_paste_behavior,
-            saved_keybindings: keybindings,
-            theme_options,
-            focus_handle: cx.focus_handle(),
-            startup_dropdown_open: false,
-            theme_dropdown_open: false,
-            image_dropdown_open: false,
-            recording_shortcut: None,
-            shortcut_error: None,
-            status_bar_enabled: preferences.status_bar.enabled,
-            status_bar_show_word_count: preferences.status_bar.show_word_count,
-            status_bar_show_cursor_position: preferences.status_bar.show_cursor_position,
-            status_bar_show_sidebar_toggle: preferences.status_bar.show_sidebar_toggle,
-            status_bar_show_mode_switch: preferences.status_bar.show_mode_switch,
-            saved_status_bar_enabled: preferences.status_bar.enabled,
-            saved_status_bar_show_word_count: preferences.status_bar.show_word_count,
-            saved_status_bar_show_cursor_position: preferences.status_bar.show_cursor_position,
-            saved_status_bar_show_sidebar_toggle: preferences.status_bar.show_sidebar_toggle,
-            saved_status_bar_show_mode_switch: preferences.status_bar.show_mode_switch,
-        }
-    }
-
-    fn selected_theme_name(&self) -> String {
-        self.theme_options
-            .iter()
-            .find(|entry| entry.id == self.selected_theme_id)
-            .map(|entry| entry.name.clone())
-            .unwrap_or_else(|| "Velotype".into())
-    }
-
-    fn has_unsaved_changes(&self) -> bool {
-        self.startup_open != self.saved_startup_open
-            || self.selected_theme_id != self.saved_theme_id
-            || self.image_paste_behavior != self.saved_image_paste_behavior
-            || normalize_shortcut_config(&self.keybindings)
-                != normalize_shortcut_config(&self.saved_keybindings)
-            || self.status_bar_enabled != self.saved_status_bar_enabled
-            || self.status_bar_show_word_count != self.saved_status_bar_show_word_count
-            || self.status_bar_show_cursor_position != self.saved_status_bar_show_cursor_position
-            || self.status_bar_show_sidebar_toggle != self.saved_status_bar_show_sidebar_toggle
-            || self.status_bar_show_mode_switch != self.saved_status_bar_show_mode_switch
-    }
-
-    fn set_nav_file(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.nav = PreferencesNav::File;
-        self.startup_dropdown_open = false;
-        self.theme_dropdown_open = false;
-        self.image_dropdown_open = false;
-        self.recording_shortcut = None;
-        cx.notify();
-    }
-
-    fn set_nav_theme(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.nav = PreferencesNav::Theme;
-        self.startup_dropdown_open = false;
-        self.theme_dropdown_open = false;
-        self.image_dropdown_open = false;
-        self.recording_shortcut = None;
-        cx.notify();
-    }
-
-    fn set_nav_image(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.nav = PreferencesNav::Image;
-        self.startup_dropdown_open = false;
-        self.theme_dropdown_open = false;
-        self.image_dropdown_open = false;
-        self.recording_shortcut = None;
-        cx.notify();
-    }
-
-    fn set_nav_shortcuts(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.nav = PreferencesNav::Shortcuts;
-        self.startup_dropdown_open = false;
-        self.theme_dropdown_open = false;
-        self.image_dropdown_open = false;
-        self.shortcut_error = None;
-        cx.notify();
-    }
-
-    fn set_nav_status_bar(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.nav = PreferencesNav::StatusBar;
-        self.startup_dropdown_open = false;
-        self.theme_dropdown_open = false;
-        self.image_dropdown_open = false;
-        self.recording_shortcut = None;
-        cx.notify();
-    }
-
-    fn toggle_startup_dropdown(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.startup_dropdown_open = !self.startup_dropdown_open;
-        self.theme_dropdown_open = false;
-        self.image_dropdown_open = false;
-        cx.notify();
-    }
-
-    fn toggle_theme_dropdown(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.theme_dropdown_open = !self.theme_dropdown_open;
-        self.startup_dropdown_open = false;
-        self.image_dropdown_open = false;
-        cx.notify();
-    }
-
-    fn toggle_image_dropdown(&mut self, _: &ClickEvent, _: &mut Window, cx: &mut Context<Self>) {
-        self.image_dropdown_open = !self.image_dropdown_open;
-        self.startup_dropdown_open = false;
-        self.theme_dropdown_open = false;
-        cx.notify();
-    }
-
-    fn cancel(&mut self, _: &ClickEvent, window: &mut Window, _: &mut Context<Self>) {
-        window.remove_window();
-    }
-
-    fn on_titlebar_close(
-        &mut self,
-        event: &ClickEvent,
-        window: &mut Window,
-        _cx: &mut Context<Self>,
-    ) {
-        if event.standard_click() {
-            window.remove_window();
-        }
-    }
-
-    fn save(&mut self, _: &ClickEvent, window: &mut Window, cx: &mut Context<Self>) {
-        if !self.has_unsaved_changes() {
-            return;
-        }
-
-        let preferences = match save_preferences_from_window(
-            self.startup_open,
-            &self.selected_theme_id,
-            self.image_paste_behavior,
-            self.keybindings.clone(),
-            &StatusBarPreferences {
-                enabled: self.status_bar_enabled,
-                show_word_count: self.status_bar_show_word_count,
-                show_cursor_position: self.status_bar_show_cursor_position,
-                show_sidebar_toggle: self.status_bar_show_sidebar_toggle,
-                show_mode_switch: self.status_bar_show_mode_switch,
-                custom_buttons: Vec::new(),
-            },
-        ) {
-            Ok(preferences) => preferences,
-            Err(err) => {
-                let strings = cx.global::<I18nManager>().strings().clone();
-                let ok = strings.info_dialog_ok;
-                let buttons = [ok.as_str()];
-                let _ = window.prompt(
-                    PromptLevel::Critical,
-                    &strings.preferences_save_failed_title,
-                    Some(&err.to_string()),
-                    &buttons,
-                    cx,
-                );
-                return;
-            }
-        };
-
-        self.apply_saved_preferences(preferences, window, cx);
-    }
-
-    fn apply_saved_preferences(
-        &mut self,
-        preferences: AppPreferences,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let theme_changed = cx.update_global::<ThemeManager, _>(|theme_manager, _cx| {
-            theme_manager.set_theme_by_id(&preferences.default_theme_id)
-        });
-        if !theme_changed {
-            let _ = cx.update_global::<ThemeManager, _>(|theme_manager, _cx| {
-                theme_manager.set_theme_by_id(DEFAULT_THEME_ID)
-            });
-        }
-        cx.clear_key_bindings();
-        install_keybindings(cx, &preferences.keybindings);
-        cx.update_global::<EditorSettings, _>(|settings, _cx| {
-            settings.status_bar_settings.status_bar_enabled = preferences.status_bar.enabled;
-            settings.status_bar_settings.status_bar_show_word_count =
-                preferences.status_bar.show_word_count;
-            settings.status_bar_settings.status_bar_show_cursor_position =
-                preferences.status_bar.show_cursor_position;
-            settings.status_bar_settings.status_bar_show_sidebar_toggle =
-                preferences.status_bar.show_sidebar_toggle;
-            settings.status_bar_settings.status_bar_show_mode_switch =
-                preferences.status_bar.show_mode_switch;
-        });
-        cx.refresh_windows();
-        window.activate_window();
-        self.focus_handle.focus(window);
-        self.saved_startup_open = self.startup_open;
-        self.saved_theme_id = self.selected_theme_id.clone();
-        self.saved_image_paste_behavior = self.image_paste_behavior;
-        self.saved_keybindings = normalize_shortcut_config(&self.keybindings);
-        self.saved_status_bar_enabled = self.status_bar_enabled;
-        self.saved_status_bar_show_word_count = self.status_bar_show_word_count;
-        self.saved_status_bar_show_cursor_position = self.status_bar_show_cursor_position;
-        self.saved_status_bar_show_sidebar_toggle = self.status_bar_show_sidebar_toggle;
-        self.saved_status_bar_show_mode_switch = self.status_bar_show_mode_switch;
-        cx.notify();
-    }
-
-    fn nav_button(
-        &self,
-        id: &'static str,
-        label: String,
-        selected: bool,
-        theme: &Theme,
-        on_click: fn(&mut Self, &ClickEvent, &mut Window, &mut Context<Self>),
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        div()
-            .h(px(34.0))
-            .w(px(156.0))
-            .px(px(12.0))
-            .flex()
-            .items_center()
-            .justify_end()
-            .rounded(px(d.menu_item_radius))
-            .cursor_pointer()
-            .text_size(px(t.dialog_body_size))
-            .font_weight(t.dialog_button_weight.to_font_weight())
-            .text_color(if selected {
-                c.dialog_primary_button_text
-            } else {
-                c.dialog_body
-            })
-            .bg(if selected {
-                c.dialog_primary_button_bg
-            } else {
-                c.dialog_secondary_button_bg
-            })
-            .hover(move |this| {
-                this.bg(if selected {
-                    c.dialog_primary_button_hover
-                } else {
-                    c.dialog_secondary_button_hover
-                })
-            })
-            .id(id)
-            .child(label)
-            .on_click(cx.listener(on_click))
-    }
-
-    fn dropdown_button(
-        id: &'static str,
-        label: String,
-        theme: &Theme,
-        on_click: fn(&mut Self, &ClickEvent, &mut Window, &mut Context<Self>),
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        div()
-            .w(px(280.0))
-            .min_h(px(36.0))
-            .px(px(12.0))
-            .flex()
-            .items_center()
-            .justify_between()
-            .rounded(px(d.menu_item_radius))
-            .border(px(d.dialog_border_width))
-            .border_color(c.dialog_border)
-            .bg(c.dialog_secondary_button_bg)
-            .hover(|this| this.bg(c.dialog_secondary_button_hover))
-            .cursor_pointer()
-            .text_size(px(t.dialog_body_size))
-            .text_color(c.dialog_body)
-            .id(id)
-            .child(label)
-            .child("v")
-            .on_click(cx.listener(on_click))
-    }
-
-    fn dropdown_item(
-        id: impl Into<ElementId>,
-        label: String,
-        selected: bool,
-        theme: &Theme,
-        on_click: impl Fn(&mut Self, &ClickEvent, &mut Window, &mut Context<Self>) + 'static,
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        div()
-            .w(px(280.0))
-            .min_h(px(30.0))
-            .px(px(12.0))
-            .flex()
-            .items_center()
-            .rounded(px(d.menu_item_radius))
-            .cursor_pointer()
-            .bg(if selected {
-                c.selection
-            } else {
-                c.dialog_surface
-            })
-            .hover(|this| this.bg(c.dialog_secondary_button_hover))
-            .text_size(px(t.dialog_body_size))
-            .text_color(c.dialog_body)
-            .id(id)
-            .child(label)
-            .on_click(cx.listener(on_click))
-    }
-
-    fn labeled_row(&self, label: &str, control: impl IntoElement, theme: &Theme) -> Div {
-        let c = &theme.colors;
-        let t = &theme.typography;
-        div()
-            .flex()
-            .flex_col()
-            .items_center()
-            .gap(px(8.0))
-            .child(
-                div()
-                    .w(px(280.0))
-                    .text_size(px(t.dialog_body_size))
-                    .font_weight(t.dialog_button_weight.to_font_weight())
-                    .text_color(c.dialog_title)
-                    .child(SharedString::from(label.to_string())),
-            )
-            .child(control)
-    }
-
-    fn render_startup_page(
-        &self,
-        theme: &Theme,
-        strings: &crate::i18n::I18nStrings,
-        cx: &mut Context<Self>,
-    ) -> Div {
-        let selected = match self.startup_open {
-            StartupOpenPreference::NewFile => strings.preferences_startup_new_file.clone(),
-            StartupOpenPreference::LastOpenedFile => {
-                strings.preferences_startup_last_opened_file.clone()
-            }
-        };
-        let mut dropdown = div()
-            .flex()
-            .flex_col()
-            .gap(px(4.0))
-            .child(Self::dropdown_button(
-                "preferences-startup-dropdown",
-                selected,
-                theme,
-                Self::toggle_startup_dropdown,
-                cx,
-            ));
-        if self.startup_dropdown_open {
-            let new_file_label = strings.preferences_startup_new_file.clone();
-            let last_file_label = strings.preferences_startup_last_opened_file.clone();
-            dropdown = dropdown
-                .child(Self::dropdown_item(
-                    "preferences-startup-new-file",
-                    new_file_label,
-                    self.startup_open == StartupOpenPreference::NewFile,
-                    theme,
-                    |this, _, _, cx| {
-                        this.startup_open = StartupOpenPreference::NewFile;
-                        this.startup_dropdown_open = false;
-                        cx.notify();
-                    },
-                    cx,
-                ))
-                .child(Self::dropdown_item(
-                    "preferences-startup-last-opened-file",
-                    last_file_label,
-                    self.startup_open == StartupOpenPreference::LastOpenedFile,
-                    theme,
-                    |this, _, _, cx| {
-                        this.startup_open = StartupOpenPreference::LastOpenedFile;
-                        this.startup_dropdown_open = false;
-                        cx.notify();
-                    },
-                    cx,
-                ));
-        }
-        self.labeled_row(&strings.preferences_startup_option, dropdown, theme)
-    }
-
-    fn render_theme_page(
-        &self,
-        theme: &Theme,
-        strings: &crate::i18n::I18nStrings,
-        cx: &mut Context<Self>,
-    ) -> Div {
-        let mut dropdown = div()
-            .flex()
-            .flex_col()
-            .gap(px(4.0))
-            .child(Self::dropdown_button(
-                "preferences-theme-dropdown",
-                self.selected_theme_name(),
-                theme,
-                Self::toggle_theme_dropdown,
-                cx,
-            ));
-        if self.theme_dropdown_open {
-            let mut list = div()
-                .id("preferences-theme-dropdown-list")
-                .flex()
-                .flex_col()
-                .gap(px(4.0))
-                .max_h(px(240.0))
-                .overflow_y_scroll();
-
-            for (index, entry) in self.theme_options.clone().into_iter().enumerate() {
-                let selected = entry.id == self.selected_theme_id;
-                list = list.child(Self::dropdown_item(
-                    ("preferences-theme-option", index),
-                    entry.name,
-                    selected,
-                    theme,
-                    move |this, _, _, cx| {
-                        this.selected_theme_id = entry.id.clone();
-                        this.theme_dropdown_open = false;
-                        cx.notify();
-                    },
-                    cx,
-                ));
-            }
-            dropdown = dropdown.child(list);
-        }
-        self.labeled_row(&strings.preferences_local_theme, dropdown, theme)
-    }
-
-    fn image_paste_behavior_label(
-        behavior: ImagePasteBehavior,
-        strings: &crate::i18n::I18nStrings,
-    ) -> String {
-        match behavior {
-            ImagePasteBehavior::None => strings.preferences_image_paste_none.clone(),
-            ImagePasteBehavior::CopyToDocumentFolder => strings
-                .preferences_image_paste_copy_to_document_folder
-                .clone(),
-            ImagePasteBehavior::CopyToAssetsFolder => strings
-                .preferences_image_paste_copy_to_assets_folder
-                .clone(),
-            ImagePasteBehavior::CopyToNamedAssetsFolder => strings
-                .preferences_image_paste_copy_to_named_assets_folder
-                .clone(),
-        }
-    }
-
-    fn render_image_page(
-        &self,
-        theme: &Theme,
-        strings: &crate::i18n::I18nStrings,
-        cx: &mut Context<Self>,
-    ) -> Div {
-        let options = [
-            ImagePasteBehavior::None,
-            ImagePasteBehavior::CopyToDocumentFolder,
-            ImagePasteBehavior::CopyToAssetsFolder,
-            ImagePasteBehavior::CopyToNamedAssetsFolder,
-        ];
-        let mut dropdown = div()
-            .flex()
-            .flex_col()
-            .gap(px(4.0))
-            .child(Self::dropdown_button(
-                "preferences-image-dropdown",
-                Self::image_paste_behavior_label(self.image_paste_behavior, strings),
-                theme,
-                Self::toggle_image_dropdown,
-                cx,
-            ));
-        if self.image_dropdown_open {
-            for (index, behavior) in options.into_iter().enumerate() {
-                let selected = behavior == self.image_paste_behavior;
-                let label = Self::image_paste_behavior_label(behavior, strings);
-                dropdown = dropdown.child(Self::dropdown_item(
-                    ("preferences-image-option", index),
-                    label,
-                    selected,
-                    theme,
-                    move |this, _, _, cx| {
-                        this.image_paste_behavior = behavior;
-                        this.image_dropdown_open = false;
-                        cx.notify();
-                    },
-                    cx,
-                ));
-            }
-        }
-        self.labeled_row(&strings.preferences_image_insert_behavior, dropdown, theme)
-    }
-
-    fn shortcut_category_label(
-        category: ShortcutCategory,
-        strings: &crate::i18n::I18nStrings,
-    ) -> String {
-        match category {
-            ShortcutCategory::File => strings.preferences_shortcuts_group_file.clone(),
-            ShortcutCategory::Edit => strings.preferences_shortcuts_group_edit.clone(),
-            ShortcutCategory::Navigation => strings.preferences_shortcuts_group_navigation.clone(),
-            ShortcutCategory::Formatting => strings.preferences_shortcuts_group_formatting.clone(),
-            ShortcutCategory::Block => strings.preferences_shortcuts_group_block.clone(),
-            ShortcutCategory::Other => strings.preferences_shortcuts_group_other.clone(),
-        }
-    }
-
-    fn shortcut_command_label(
-        command: ShortcutCommand,
-        strings: &crate::i18n::I18nStrings,
-    ) -> String {
-        match command {
-            ShortcutCommand::Newline => strings.preferences_shortcut_newline.clone(),
-            ShortcutCommand::DeleteBack => strings.preferences_shortcut_delete_back.clone(),
-            ShortcutCommand::Delete => strings.preferences_shortcut_delete.clone(),
-            ShortcutCommand::WordDeleteBack => {
-                strings.preferences_shortcut_word_delete_back.clone()
-            }
-            ShortcutCommand::WordDeleteForward => {
-                strings.preferences_shortcut_word_delete_forward.clone()
-            }
-            ShortcutCommand::FocusPrev => strings.preferences_shortcut_focus_prev.clone(),
-            ShortcutCommand::FocusNext => strings.preferences_shortcut_focus_next.clone(),
-            ShortcutCommand::MoveLeft => strings.preferences_shortcut_move_left.clone(),
-            ShortcutCommand::MoveRight => strings.preferences_shortcut_move_right.clone(),
-            ShortcutCommand::WordMoveLeft => strings.preferences_shortcut_word_move_left.clone(),
-            ShortcutCommand::WordMoveRight => strings.preferences_shortcut_word_move_right.clone(),
-            ShortcutCommand::Home => strings.preferences_shortcut_home.clone(),
-            ShortcutCommand::End => strings.preferences_shortcut_end.clone(),
-            ShortcutCommand::BlockUp => strings.preferences_shortcut_block_up.clone(),
-            ShortcutCommand::BlockDown => strings.preferences_shortcut_block_down.clone(),
-            ShortcutCommand::PageUp => strings.preferences_shortcut_page_up.clone(),
-            ShortcutCommand::PageDown => strings.preferences_shortcut_page_down.clone(),
-            ShortcutCommand::JumpToTop => strings.preferences_shortcut_jump_to_top.clone(),
-            ShortcutCommand::JumpToBottom => strings.preferences_shortcut_jump_to_bottom.clone(),
-            ShortcutCommand::SelectLeft => strings.preferences_shortcut_select_left.clone(),
-            ShortcutCommand::SelectRight => strings.preferences_shortcut_select_right.clone(),
-            ShortcutCommand::WordSelectLeft => {
-                strings.preferences_shortcut_word_select_left.clone()
-            }
-            ShortcutCommand::WordSelectRight => {
-                strings.preferences_shortcut_word_select_right.clone()
-            }
-            ShortcutCommand::SelectHome => strings.preferences_shortcut_select_home.clone(),
-            ShortcutCommand::SelectEnd => strings.preferences_shortcut_select_end.clone(),
-            ShortcutCommand::SelectAll => strings.preferences_shortcut_select_all.clone(),
-            ShortcutCommand::Copy => strings.preferences_shortcut_copy.clone(),
-            ShortcutCommand::Cut => strings.preferences_shortcut_cut.clone(),
-            ShortcutCommand::Paste => strings.preferences_shortcut_paste.clone(),
-            ShortcutCommand::Undo => strings.preferences_shortcut_undo.clone(),
-            ShortcutCommand::Redo => strings.preferences_shortcut_redo.clone(),
-            ShortcutCommand::BoldSelection => strings.preferences_shortcut_bold_selection.clone(),
-            ShortcutCommand::ItalicSelection => {
-                strings.preferences_shortcut_italic_selection.clone()
-            }
-            ShortcutCommand::UnderlineSelection => {
-                strings.preferences_shortcut_underline_selection.clone()
-            }
-            ShortcutCommand::CodeSelection => strings.preferences_shortcut_code_selection.clone(),
-            ShortcutCommand::IndentBlock => strings.preferences_shortcut_indent_block.clone(),
-            ShortcutCommand::OutdentBlock => strings.preferences_shortcut_outdent_block.clone(),
-            ShortcutCommand::ExitCodeBlock => strings.preferences_shortcut_exit_code_block.clone(),
-            // ShortcutCommand::SaveDocument => strings.preferences_shortcut_save_document.clone(),
-            // ShortcutCommand::SaveDocumentAs => {
-            //     strings.preferences_shortcut_save_document_as.clone()
-            // }
-            // ShortcutCommand::NewWindow => strings.preferences_shortcut_new_window.clone(),
-            // ShortcutCommand::OpenFile => strings.preferences_shortcut_open_file.clone(),
-            // ShortcutCommand::QuitApplication => {
-            //     strings.preferences_shortcut_quit_application.clone()
-            // }
-            // ShortcutCommand::CloseWindow => strings.preferences_shortcut_close_window.clone(),
-            // ShortcutCommand::DismissTransientUi => {
-            //     strings.preferences_shortcut_dismiss_transient_ui.clone()
-            // }
-            // ShortcutCommand::ToggleViewMode => {
-            //     strings.preferences_shortcut_toggle_view_mode.clone()
-            // }
-            // ShortcutCommand::ToggleWorkspace => {
-            //     strings.preferences_shortcut_toggle_workspace.clone()
-            // }
-        }
-    }
-
-    fn format_template(template: &str, key: &str, value: &str) -> String {
-        template.replace(key, value)
-    }
-
-    fn begin_recording_shortcut(
-        &mut self,
-        command: ShortcutCommand,
-        _: &ClickEvent,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.recording_shortcut = Some(command);
-        self.shortcut_error = None;
-        window.focus(&self.focus_handle);
-        cx.notify();
-    }
-
-    fn reset_shortcut(
-        &mut self,
-        command: ShortcutCommand,
-        _: &ClickEvent,
-        _: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if let Some(definition) = shortcut_definitions()
-            .iter()
-            .find(|definition| definition.command == command)
-        {
-            self.keybindings.remove(definition.id);
-        }
-        if self.recording_shortcut == Some(command) {
-            self.recording_shortcut = None;
-        }
-        self.shortcut_error = None;
-        cx.notify();
-    }
-
-    fn capture_shortcut_key(
-        &mut self,
-        event: &KeyDownEvent,
-        _: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let Some(command) = self.recording_shortcut else {
-            return;
-        };
-        cx.stop_propagation();
-        if event.is_held {
-            return;
-        }
-
-        let key = event.keystroke.unparse();
-        if key == "escape" {
-            self.recording_shortcut = None;
-            self.shortcut_error = None;
-            cx.notify();
-            return;
-        }
-
-        let Some(keys) = normalize_shortcut_keys(std::slice::from_ref(&key)) else {
-            let strings = cx.global::<I18nManager>().strings();
-            self.shortcut_error = Some(Self::format_template(
-                &strings.preferences_shortcut_invalid_template,
-                "{shortcut}",
-                &key,
-            ));
-            cx.notify();
-            return;
-        };
-
-        if let Some(conflict) = shortcut_conflict_for(command, &keys, &self.keybindings) {
-            let strings = cx.global::<I18nManager>().strings();
-            let label = Self::shortcut_command_label(conflict.command, strings);
-            self.shortcut_error = Some(Self::format_template(
-                &strings.preferences_shortcut_conflict_template,
-                "{command}",
-                &label,
-            ));
-            cx.notify();
-            return;
-        }
-
-        if let Some(definition) = shortcut_definitions()
-            .iter()
-            .find(|definition| definition.command == command)
-        {
-            let defaults = definition
-                .default_keys
-                .iter()
-                .map(|key| key.to_string())
-                .collect::<Vec<_>>();
-            if keys == defaults {
-                self.keybindings.remove(definition.id);
-            } else {
-                self.keybindings.insert(definition.id.to_string(), keys);
-            }
-        }
-        self.recording_shortcut = None;
-        self.shortcut_error = None;
-        cx.notify();
-    }
-
-    fn shortcut_chip(label: &str, theme: &Theme) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        div()
-            .min_w(px(58.0))
-            .h(px(24.0))
-            .px(px(8.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .rounded(px((d.menu_item_radius - 1.0).max(0.0)))
-            .border(px(d.dialog_border_width))
-            .border_color(c.dialog_border)
-            .bg(c.code_bg)
-            .text_size(px((t.dialog_body_size - 1.0).max(10.0)))
-            .text_color(c.code_text)
-            .child(SharedString::from(label.to_string()))
-    }
-
-    fn shortcut_action_button(
-        id: impl Into<ElementId>,
-        label: String,
-        theme: &Theme,
-        on_click: impl Fn(&mut Self, &ClickEvent, &mut Window, &mut Context<Self>) + 'static,
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        div()
-            .id(id)
-            .h(px(28.0))
-            .px(px(10.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .rounded(px((d.dialog_radius - 5.0).max(0.0)))
-            .border(px(d.dialog_border_width))
-            .border_color(c.dialog_border)
-            .bg(c.dialog_secondary_button_bg)
-            .hover(|this| this.bg(c.dialog_secondary_button_hover))
-            .cursor_pointer()
-            .text_size(px((t.dialog_button_size - 1.0).max(10.0)))
-            .font_weight(t.dialog_button_weight.to_font_weight())
-            .text_color(c.dialog_secondary_button_text)
-            .child(label)
-            .on_click(cx.listener(on_click))
-    }
-
-    fn render_shortcut_row(
-        &self,
-        definition: ShortcutDefinition,
-        theme: &Theme,
-        strings: &crate::i18n::I18nStrings,
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let is_recording = self.recording_shortcut == Some(definition.command);
-        let keys = resolved_shortcut_keys(&self.keybindings, definition.command);
-        let label = Self::shortcut_command_label(definition.command, strings);
-        let command = definition.command;
-
-        let mut chips = div().flex().flex_wrap().gap(px(6.0));
-        if is_recording {
-            chips = chips.child(Self::shortcut_chip(
-                &strings.preferences_shortcut_recording,
-                theme,
-            ));
-        } else {
-            for key in keys {
-                chips = chips.child(Self::shortcut_chip(&key, theme));
-            }
-        }
-
-        div()
-            .w_full()
-            .min_h(px(42.0))
-            .px(px(10.0))
-            .py(px(6.0))
-            .flex()
-            .items_center()
-            .justify_between()
-            .gap(px(12.0))
-            .rounded(px(d.menu_item_radius))
-            .bg(c.dialog_surface)
-            .child(
-                div()
-                    .min_w(px(144.0))
-                    .text_size(px(t.dialog_body_size))
-                    .text_color(c.dialog_body)
-                    .child(label),
-            )
-            .child(div().flex_1().child(chips))
-            .child(
-                div()
-                    .flex()
-                    .gap(px(6.0))
-                    .child(Self::shortcut_action_button(
-                        ("preferences-shortcut-record", definition.command as u32),
-                        strings.preferences_shortcut_record.clone(),
-                        theme,
-                        move |this, event, window, cx| {
-                            this.begin_recording_shortcut(command, event, window, cx)
-                        },
-                        cx,
-                    ))
-                    .child(Self::shortcut_action_button(
-                        ("preferences-shortcut-reset", definition.command as u32),
-                        strings.preferences_shortcut_reset.clone(),
-                        theme,
-                        move |this, event, window, cx| {
-                            this.reset_shortcut(command, event, window, cx)
-                        },
-                        cx,
-                    )),
-            )
-    }
-
-    fn render_shortcuts_page(
-        &self,
-        theme: &Theme,
-        strings: &crate::i18n::I18nStrings,
-        cx: &mut Context<Self>,
-    ) -> Div {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let mut content = div()
-            .id("preferences-shortcuts-scroll")
-            .w_full()
-            .flex_1()
-            .min_h(px(0.0))
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .gap(px(18.0))
-            .pr(px(4.0));
-
-        let categories = [
-            ShortcutCategory::File,
-            ShortcutCategory::Edit,
-            ShortcutCategory::Navigation,
-            ShortcutCategory::Formatting,
-            ShortcutCategory::Block,
-            ShortcutCategory::Other,
-        ];
-
-        for category in categories {
-            let mut group = div().w_full().flex().flex_col().gap(px(8.0)).child(
-                div()
-                    .w_full()
-                    .flex()
-                    .items_center()
-                    .gap(px(10.0))
-                    .child(
-                        div()
-                            .text_size(px(t.dialog_body_size))
-                            .font_weight(t.dialog_button_weight.to_font_weight())
-                            .text_color(c.dialog_title)
-                            .child(Self::shortcut_category_label(category, strings)),
-                    )
-                    .child(
-                        div()
-                            .flex_1()
-                            .h(px(d.dialog_border_width.max(1.0)))
-                            .bg(c.dialog_border),
-                    ),
-            );
-            for definition in shortcut_definitions()
-                .iter()
-                .copied()
-                .filter(|definition| definition.category == category)
-            {
-                group = group.child(self.render_shortcut_row(definition, theme, strings, cx));
-            }
-            content = content.child(group);
-        }
-
-        let mut page = div()
-            .w_full()
-            .h_full()
-            .flex_1()
-            .min_h(px(0.0))
-            .flex()
-            .flex_col()
-            .items_center()
-            .gap(px(8.0));
-        if let Some(error) = &self.shortcut_error {
-            page = page.child(
-                div()
-                    .w_full()
-                    .flex_shrink_0()
-                    .text_size(px(t.dialog_body_size))
-                    .text_color(c.dialog_danger_button_bg)
-                    .child(error.clone()),
-            );
-        }
-        page.child(content)
-    }
-
-    fn render_status_bar_page(
-        &self,
-        theme: &Theme,
-        strings: &crate::i18n::I18nStrings,
-        cx: &mut Context<Self>,
-    ) -> Div {
-        let c = &theme.colors;
-        let t = &theme.typography;
-
-        let switch_row = |label: &str,
-                          checked: bool,
-                          on_click: fn(&mut Self, &ClickEvent, &mut Window, &mut Context<Self>),
-                          cx: &mut Context<Self>| {
-            div()
-                .w(px(280.0))
-                .flex()
-                .items_center()
-                .justify_between()
-                .child(
-                    div()
-                        .text_size(px(t.dialog_body_size))
-                        .text_color(c.dialog_body)
-                        .child(SharedString::from(label.to_string())),
-                )
-                .child(
-                    Switch::new(ElementId::Name(
-                        format!("preferences-toggle-{}", label).into(),
-                    ))
-                    .checked(checked)
-                    .on_click(cx.listener(on_click)),
-                )
-        };
-
-        let items = div()
-            .flex()
-            .flex_col()
-            .gap(px(12.0))
-            .child(switch_row(
-                &strings.preferences_status_bar_enabled,
-                self.status_bar_enabled,
-                |this, _, _, cx| {
-                    this.status_bar_enabled = !this.status_bar_enabled;
-                    cx.notify();
-                },
-                cx,
-            ))
-            .child(switch_row(
-                &strings.preferences_status_bar_show_word_count,
-                self.status_bar_show_word_count,
-                |this, _, _, cx| {
-                    this.status_bar_show_word_count = !this.status_bar_show_word_count;
-                    cx.notify();
-                },
-                cx,
-            ))
-            .child(switch_row(
-                &strings.preferences_status_bar_show_cursor_position,
-                self.status_bar_show_cursor_position,
-                |this, _, _, cx| {
-                    this.status_bar_show_cursor_position = !this.status_bar_show_cursor_position;
-                    cx.notify();
-                },
-                cx,
-            ))
-            .child(switch_row(
-                &strings.preferences_status_bar_show_sidebar_toggle,
-                self.status_bar_show_sidebar_toggle,
-                |this, _, _, cx| {
-                    this.status_bar_show_sidebar_toggle = !this.status_bar_show_sidebar_toggle;
-                    cx.notify();
-                },
-                cx,
-            ))
-            .child(switch_row(
-                &strings.preferences_status_bar_show_mode_switch,
-                self.status_bar_show_mode_switch,
-                |this, _, _, cx| {
-                    this.status_bar_show_mode_switch = !this.status_bar_show_mode_switch;
-                    cx.notify();
-                },
-                cx,
-            ));
-
-        div()
-            .w_full()
-            .flex_1()
-            .min_h(px(0.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .child(items)
-    }
-}
-
-impl Render for PreferencesWindow {
-    fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let theme = cx.global::<ThemeManager>().current().clone();
-        let strings = cx.global::<I18nManager>().strings().clone();
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let can_save = self.has_unsaved_changes();
-        let window_title =
-            SharedString::from(format!("Velotype - {}", strings.preferences_window_title));
-        window.set_window_title(window_title.as_ref());
-        let titlebar_height = custom_titlebar_height(window, d);
-
-        let content = div()
-            .size_full()
-            .pt(px(titlebar_height))
-            .flex()
-            .key_context("Preferences")
-            .track_focus(&self.focus_handle)
-            .on_key_down(cx.listener(Self::capture_shortcut_key))
-            .bg(c.editor_background)
-            .text_color(c.dialog_body)
-            .child(
-                div()
-                    .w(relative(0.3))
-                    .h_full()
-                    .pr(px(20.0))
-                    .flex()
-                    .items_center()
-                    .justify_end()
-                    .border_r(px(d.dialog_border_width))
-                    .border_color(c.dialog_border)
-                    .child(
-                        div()
-                            .flex()
-                            .flex_col()
-                            .gap(px(8.0))
-                            .child(self.nav_button(
-                                "preferences-nav-file",
-                                strings.preferences_nav_file.clone(),
-                                self.nav == PreferencesNav::File,
-                                &theme,
-                                Self::set_nav_file,
-                                cx,
-                            ))
-                            .child(self.nav_button(
-                                "preferences-nav-theme",
-                                strings.preferences_nav_theme.clone(),
-                                self.nav == PreferencesNav::Theme,
-                                &theme,
-                                Self::set_nav_theme,
-                                cx,
-                            ))
-                            .child(self.nav_button(
-                                "preferences-nav-image",
-                                strings.preferences_nav_image.clone(),
-                                self.nav == PreferencesNav::Image,
-                                &theme,
-                                Self::set_nav_image,
-                                cx,
-                            ))
-                            .child(self.nav_button(
-                                "preferences-nav-shortcuts",
-                                strings.preferences_nav_shortcuts.clone(),
-                                self.nav == PreferencesNav::Shortcuts,
-                                &theme,
-                                Self::set_nav_shortcuts,
-                                cx,
-                            ))
-                            .child(self.nav_button(
-                                "preferences-nav-status-bar",
-                                strings.preferences_nav_status_bar.clone(),
-                                self.nav == PreferencesNav::StatusBar,
-                                &theme,
-                                Self::set_nav_status_bar,
-                                cx,
-                            )),
-                    ),
-            )
-            .child(
-                div()
-                    .w(relative(0.7))
-                    .h_full()
-                    .p(px(d.dialog_padding))
-                    .flex()
-                    .flex_col()
-                    .gap(px(d.dialog_gap))
-                    .child(
-                        div()
-                            .w_full()
-                            .flex_1()
-                            .min_h(px(0.0))
-                            .flex()
-                            .flex_col()
-                            .items_center()
-                            .gap(px(d.dialog_gap * 1.5))
-                            .child(
-                                div()
-                                    .flex_shrink_0()
-                                    .text_size(px(t.dialog_title_size))
-                                    .font_weight(t.dialog_title_weight.to_font_weight())
-                                    .text_color(c.dialog_title)
-                                    .child(match self.nav {
-                                        PreferencesNav::File => {
-                                            strings.preferences_nav_file.clone()
-                                        }
-                                        PreferencesNav::Theme => {
-                                            strings.preferences_nav_theme.clone()
-                                        }
-                                        PreferencesNav::Image => {
-                                            strings.preferences_nav_image.clone()
-                                        }
-                                        PreferencesNav::Shortcuts => {
-                                            strings.preferences_nav_shortcuts.clone()
-                                        }
-                                        PreferencesNav::StatusBar => {
-                                            strings.preferences_nav_status_bar.clone()
-                                        }
-                                    }),
-                            )
-                            .child(match self.nav {
-                                PreferencesNav::File => div()
-                                    .w_full()
-                                    .flex_1()
-                                    .min_h(px(0.0))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .child(self.render_startup_page(&theme, &strings, cx))
-                                    .into_any_element(),
-                                PreferencesNav::Theme => div()
-                                    .w_full()
-                                    .flex_1()
-                                    .min_h(px(0.0))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .child(self.render_theme_page(&theme, &strings, cx))
-                                    .into_any_element(),
-                                PreferencesNav::Image => div()
-                                    .w_full()
-                                    .flex_1()
-                                    .min_h(px(0.0))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .child(self.render_image_page(&theme, &strings, cx))
-                                    .into_any_element(),
-                                PreferencesNav::Shortcuts => div()
-                                    .w_full()
-                                    .flex_1()
-                                    .min_h(px(0.0))
-                                    .child(self.render_shortcuts_page(&theme, &strings, cx))
-                                    .into_any_element(),
-                                PreferencesNav::StatusBar => div()
-                                    .w_full()
-                                    .flex_1()
-                                    .min_h(px(0.0))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .child(self.render_status_bar_page(&theme, &strings, cx))
-                                    .into_any_element(),
-                            }),
-                    )
-                    .child(
-                        div()
-                            .w_full()
-                            .flex_shrink_0()
-                            .flex()
-                            .justify_end()
-                            .gap(px(d.dialog_button_gap))
-                            .child(
-                                div()
-                                    .id("preferences-cancel")
-                                    .h(px(d.dialog_button_height))
-                                    .px(px(d.dialog_button_padding_x))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .rounded(px((d.dialog_radius - 4.0).max(0.0)))
-                                    .border(px(d.dialog_border_width))
-                                    .border_color(c.dialog_border)
-                                    .bg(c.dialog_secondary_button_bg)
-                                    .hover(|this| this.bg(c.dialog_secondary_button_hover))
-                                    .cursor_pointer()
-                                    .text_size(px(t.dialog_button_size))
-                                    .font_weight(t.dialog_button_weight.to_font_weight())
-                                    .text_color(c.dialog_secondary_button_text)
-                                    .child(strings.preferences_cancel.clone())
-                                    .on_click(cx.listener(Self::cancel)),
-                            )
-                            .child(
-                                div()
-                                    .id("preferences-save")
-                                    .h(px(d.dialog_button_height))
-                                    .px(px(d.dialog_button_padding_x))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .rounded(px((d.dialog_radius - 4.0).max(0.0)))
-                                    .border(px(if can_save { 0.0 } else { d.dialog_border_width }))
-                                    .border_color(c.dialog_border)
-                                    .bg(if can_save {
-                                        c.dialog_primary_button_bg
-                                    } else {
-                                        c.dialog_secondary_button_bg
-                                    })
-                                    .hover(move |this| {
-                                        if can_save {
-                                            this.bg(c.dialog_primary_button_hover)
-                                        } else {
-                                            this.bg(c.dialog_secondary_button_bg)
-                                        }
-                                    })
-                                    .when(can_save, |this| this.cursor_pointer())
-                                    .text_size(px(t.dialog_button_size))
-                                    .font_weight(t.dialog_button_weight.to_font_weight())
-                                    .text_color(if can_save {
-                                        c.dialog_primary_button_text
-                                    } else {
-                                        c.dialog_secondary_button_text
-                                    })
-                                    .child(strings.preferences_save.clone())
-                                    .on_click(cx.listener(Self::save)),
-                            ),
-                    ),
-            );
-
-        let root = div()
-            .size_full()
-            .relative()
-            .bg(c.editor_background)
-            .child(content);
-
-        if let Some(titlebar) = render_custom_titlebar(
-            "preferences-titlebar",
-            window_title,
-            &theme,
-            window,
-            cx,
-            Self::on_titlebar_close,
-        ) {
-            root.child(titlebar)
-        } else {
-            root
-        }
-    }
-}
-
-fn open_preferences_window_with_state(
-    cx: &mut App,
-    preferences: AppPreferences,
-    theme_options: Vec<ThemeCatalogEntry>,
-    title: String,
-) -> WindowHandle<PreferencesWindow> {
-    let bounds = Bounds::centered(None, size(px(720.0), px(480.0)), cx);
-    let window_title = SharedString::from(format!("Velotype - {title}"));
-    let handle = cx
-        .open_window(
-            velotype_window_options(window_title, bounds),
-            move |_window, cx| {
-                cx.new(move |cx| PreferencesWindow::new(preferences, theme_options, cx))
-            },
-        )
-        .expect("preferences window should open");
-
-    handle
-        .update(cx, |preferences, window, _cx| {
-            window.activate_window();
-            preferences.focus_handle.focus(window);
-        })
-        .expect("newly opened preferences window should be updateable");
-
-    handle
-}
-
-pub(crate) fn open_preferences_window(cx: &mut App) -> WindowHandle<PreferencesWindow> {
-    let preferences = match read_app_preferences() {
-        Ok(preferences) => preferences,
-        Err(err) => {
-            eprintln!("failed to read app preferences: {err}");
-            AppPreferences::default()
-        }
-    };
-    let theme_options = cx.global::<ThemeManager>().available_themes().to_vec();
-    let title = cx
-        .global::<I18nManager>()
-        .strings()
-        .preferences_window_title
-        .clone();
-    open_preferences_window_with_state(cx, preferences, theme_options, title)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        AppPreferences, EditorSettings, ImagePasteBehavior, StartupOpenPreference,
-        StatusBarPreferences, load_or_create_app_preferences_with_dirs_and_locales,
-        open_preferences_window_with_state, read_app_preferences_with_dirs,
-        save_app_preferences_with_dirs, save_preferences_from_window_with_dirs,
-    };
-    use crate::config::VelotypeConfigDirs;
-    use crate::i18n::I18nManager;
-    use crate::theme::{ThemeCatalogEntry, ThemeManager};
-    use gpui::TestAppContext;
-    use std::collections::BTreeMap;
-
-    fn init_preferences_test_app(cx: &mut TestAppContext) {
-        cx.update(|cx| {
-            I18nManager::init_with_language_id(cx, "en-US");
-            ThemeManager::init_with_theme_id(cx, "velotype");
-            crate::components::init(cx);
-            EditorSettings::init(cx, true);
-        });
-    }
-
-    fn default_theme_options() -> Vec<ThemeCatalogEntry> {
-        vec![ThemeCatalogEntry {
-            id: "velotype".into(),
-            name: "Velotype".into(),
-        }]
-    }
-
-    #[test]
-    fn missing_preferences_file_returns_defaults() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-missing-{}",
-            uuid::Uuid::new_v4()
-        ));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let preferences =
-            read_app_preferences_with_dirs(&dirs).expect("missing preferences should load");
-        assert_eq!(preferences, AppPreferences::default());
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn partial_or_invalid_preferences_fall_back_by_field() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-partial-{}",
-            uuid::Uuid::new_v4()
-        ));
-        std::fs::create_dir_all(&root).expect("temp root should exist");
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        std::fs::write(
-            dirs.app_config_file(),
-            r#"
-                [startup]
-                open = "not-valid"
-
-                [theme]
-                default_theme_id = "velotype-light"
-            "#,
-        )
-        .expect("preferences should be written");
-
-        let preferences =
-            read_app_preferences_with_dirs(&dirs).expect("partial preferences should load");
-        assert_eq!(preferences.startup_open, StartupOpenPreference::NewFile);
-        assert_eq!(preferences.default_language_id, "en-US");
-        assert_eq!(preferences.default_theme_id, "velotype-light");
-        assert_eq!(preferences.image_paste_behavior, ImagePasteBehavior::None);
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn invalid_image_paste_behavior_falls_back_to_none() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-image-invalid-{}",
-            uuid::Uuid::new_v4()
-        ));
-        std::fs::create_dir_all(&root).expect("temp root should exist");
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        std::fs::write(
-            dirs.app_config_file(),
-            r#"
-                [editor]
-                image_paste_behavior = "somewhere-dangerous"
-            "#,
-        )
-        .expect("preferences should be written");
-
-        let preferences = read_app_preferences_with_dirs(&dirs).expect("preferences should load");
-        assert_eq!(preferences.image_paste_behavior, ImagePasteBehavior::None);
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn damaged_preferences_file_returns_defaults() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-damaged-{}",
-            uuid::Uuid::new_v4()
-        ));
-        std::fs::create_dir_all(&root).expect("temp root should exist");
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        std::fs::write(dirs.app_config_file(), "not = [valid")
-            .expect("preferences should be written");
-
-        let preferences =
-            read_app_preferences_with_dirs(&dirs).expect("damaged preferences should load");
-        assert_eq!(preferences, AppPreferences::default());
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn saves_and_reads_preferences() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-save-{}",
-            uuid::Uuid::new_v4()
-        ));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let preferences = AppPreferences {
-            startup_open: StartupOpenPreference::LastOpenedFile,
-            default_language_id: "zh-CN".into(),
-            default_theme_id: "velotype-light".into(),
-            show_table_headers: false,
-            image_paste_behavior: ImagePasteBehavior::CopyToAssetsFolder,
-            keybindings: BTreeMap::new(),
-            status_bar: StatusBarPreferences::default(),
-        };
-
-        save_app_preferences_with_dirs(&preferences, &dirs)
-            .expect("preferences should save to config.toml");
-        let loaded = read_app_preferences_with_dirs(&dirs).expect("preferences should read back");
-        assert_eq!(loaded, preferences);
-
-        let text =
-            std::fs::read_to_string(dirs.app_config_file()).expect("config.toml should exist");
-        assert!(text.contains("open = \"last_opened_file\""));
-        assert!(text.contains("default_language_id = \"zh-CN\""));
-        assert!(text.contains("default_theme_id = \"velotype-light\""));
-        assert!(text.contains("show_table_headers = false"));
-        assert!(text.contains("image_paste_behavior = \"copy_to_assets_folder\""));
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn missing_preferences_file_is_created_with_detected_language() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-create-{}",
-            uuid::Uuid::new_v4()
-        ));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let preferences = load_or_create_app_preferences_with_dirs_and_locales(&dirs, ["zh-HK"])
-            .expect("preferences should be created");
-        assert_eq!(preferences.default_language_id, "zh-CN");
-        assert!(dirs.app_config_file().exists());
-        let text =
-            std::fs::read_to_string(dirs.app_config_file()).expect("config.toml should exist");
-        assert!(text.contains("[language]"));
-        assert!(text.contains("default_language_id = \"zh-CN\""));
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn legacy_preferences_are_normalized_with_language() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-legacy-{}",
-            uuid::Uuid::new_v4()
-        ));
-        std::fs::create_dir_all(&root).expect("temp root should exist");
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        std::fs::write(
-            dirs.app_config_file(),
-            r#"
-                [startup]
-                open = "last_opened_file"
-
-                [theme]
-                default_theme_id = "velotype-light"
-            "#,
-        )
-        .expect("legacy preferences should be written");
-
-        let preferences = load_or_create_app_preferences_with_dirs_and_locales(&dirs, ["en-GB"])
-            .expect("legacy preferences should normalize");
-        assert_eq!(
-            preferences.startup_open,
-            StartupOpenPreference::LastOpenedFile
-        );
-        assert_eq!(preferences.default_language_id, "en-US");
-        assert_eq!(preferences.default_theme_id, "velotype-light");
-        let text =
-            std::fs::read_to_string(dirs.app_config_file()).expect("config.toml should exist");
-        assert!(text.contains("[language]"));
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn saving_preferences_window_preserves_language() {
-        let root = std::env::temp_dir().join(format!(
-            "velotype-preferences-window-{}",
-            uuid::Uuid::new_v4()
-        ));
-        let dirs = VelotypeConfigDirs::from_root(&root);
-        let preferences = AppPreferences {
-            startup_open: StartupOpenPreference::NewFile,
-            default_language_id: "zh-CN".into(),
-            default_theme_id: "velotype".into(),
-            show_table_headers: true,
-            image_paste_behavior: ImagePasteBehavior::None,
-            keybindings: BTreeMap::new(),
-            status_bar: StatusBarPreferences::default(),
-        };
-        save_app_preferences_with_dirs(&preferences, &dirs)
-            .expect("preferences should save to config.toml");
-
-        let saved = save_preferences_from_window_with_dirs(
-            StartupOpenPreference::LastOpenedFile,
-            "velotype-light",
-            ImagePasteBehavior::CopyToNamedAssetsFolder,
-            BTreeMap::from([("save_document".to_string(), vec!["ctrl-alt-s".to_string()])]),
-            &StatusBarPreferences::default(),
-            &dirs,
-        )
-        .expect("window preferences should save");
-        assert_eq!(saved.default_language_id, "zh-CN");
-        assert_eq!(saved.startup_open, StartupOpenPreference::LastOpenedFile);
-        assert_eq!(saved.default_theme_id, "velotype-light");
-        assert_eq!(
-            saved.image_paste_behavior,
-            ImagePasteBehavior::CopyToNamedAssetsFolder
-        );
-        assert_eq!(
-            saved.keybindings.get("save_document"),
-            Some(&vec!["ctrl-alt-s".to_string()])
-        );
-        let _ = std::fs::remove_dir_all(root);
-    }
-
-    #[gpui::test]
-    async fn preferences_window_activates_and_focuses_on_open(cx: &mut TestAppContext) {
-        init_preferences_test_app(cx);
-
-        let handle = cx.update(|cx| {
-            open_preferences_window_with_state(
-                cx,
-                AppPreferences::default(),
-                default_theme_options(),
-                "Preferences".into(),
-            )
-        });
-        cx.run_until_parked();
-
-        let active_window = cx.update(|cx| cx.active_window().expect("window should be active"));
-        assert_eq!(active_window.window_id(), handle.window_id());
-        assert!(
-            handle
-                .update(cx, |preferences, window, _cx| preferences
-                    .focus_handle
-                    .is_focused(window))
-                .expect("preferences window should be updateable")
-        );
-        assert!(
-            !handle
-                .update(cx, |preferences, _window, _cx| preferences
-                    .has_unsaved_changes())
-                .expect("preferences window should be updateable")
-        );
-    }
-
-    #[gpui::test]
-    async fn preferences_dirty_state_tracks_draft_changes(cx: &mut TestAppContext) {
-        init_preferences_test_app(cx);
-
-        let handle = cx.update(|cx| {
-            open_preferences_window_with_state(
-                cx,
-                AppPreferences::default(),
-                default_theme_options(),
-                "Preferences".into(),
-            )
-        });
-        cx.run_until_parked();
-
-        handle
-            .update(cx, |preferences, _window, _cx| {
-                assert!(!preferences.has_unsaved_changes());
-                preferences.startup_open = StartupOpenPreference::LastOpenedFile;
-                assert!(preferences.has_unsaved_changes());
-                preferences.startup_open = StartupOpenPreference::NewFile;
-                assert!(!preferences.has_unsaved_changes());
-
-                preferences.image_paste_behavior = ImagePasteBehavior::CopyToAssetsFolder;
-                assert!(preferences.has_unsaved_changes());
-                preferences.image_paste_behavior = ImagePasteBehavior::None;
-                assert!(!preferences.has_unsaved_changes());
-
-                preferences
-                    .keybindings
-                    .insert("save_document".into(), vec!["ctrl-alt-s".into()]);
-                assert!(preferences.has_unsaved_changes());
-            })
-            .expect("preferences window should be updateable");
-    }
-
-    #[gpui::test]
-    async fn applying_saved_preferences_keeps_window_open_and_focused(cx: &mut TestAppContext) {
-        init_preferences_test_app(cx);
-
-        let handle = cx.update(|cx| {
-            open_preferences_window_with_state(
-                cx,
-                AppPreferences::default(),
-                default_theme_options(),
-                "Preferences".into(),
-            )
-        });
-        cx.run_until_parked();
-
-        handle
-            .update(cx, |preferences, window, cx| {
-                preferences.startup_open = StartupOpenPreference::LastOpenedFile;
-                assert!(preferences.has_unsaved_changes());
-                let saved = AppPreferences {
-                    startup_open: StartupOpenPreference::LastOpenedFile,
-                    ..AppPreferences::default()
-                };
-                preferences.apply_saved_preferences(saved, window, cx);
-            })
-            .expect("preferences window should be updateable");
-        cx.run_until_parked();
-
-        assert_eq!(cx.update(|cx| cx.windows().len()), 1);
-        let active_window = cx.update(|cx| cx.active_window().expect("window should be active"));
-        assert_eq!(active_window.window_id(), handle.window_id());
-        assert!(
-            handle
-                .update(cx, |preferences, window, _cx| preferences
-                    .focus_handle
-                    .is_focused(window))
-                .expect("preferences window should remain updateable")
-        );
-        assert!(
-            !handle
-                .update(cx, |preferences, _window, _cx| preferences
-                    .has_unsaved_changes())
-                .expect("preferences window should remain updateable")
-        );
-    }
 }
