+++ crates/opennote-velotype/src/editor/render.rs
@@ -6,23 +6,14 @@ use std::time::{Duration, Instant};
 
 use gpui::*;
 
-use super::{Editor, InfoDialogKind, MountedRun, workspace::workspace_panel_width_for_viewport};
-use crate::app_menu::dispatch_menu_action_for_editor;
-use crate::components::CalloutVariant;
-use crate::components::{AddLanguageConfig, AddThemeConfig, Block, NoRecentFiles};
-use crate::i18n::{I18nManager, I18nStrings};
+use super::{Editor, MountedRun, workspace::workspace_panel_width_for_viewport};
+use crate::components::{Block, CalloutVariant};
+use crate::i18n::I18nManager;
 use crate::theme::{Theme, ThemeDimensions, ThemeManager};
-use crate::window_chrome::custom_titlebar_height;
-
-pub(crate) const ABOUT_GITHUB_URL: &str = "https://github.com/manyougz/velotype";
 
 /// Rows within this many pixels of the viewport stay mounted, so a fast flick
 /// paints them before they scroll in instead of showing a blank edge.
 const RENDER_OVERDRAW_PX: f32 = 800.0;
-
-pub(crate) fn open_about_github_url(cx: &mut App) {
-    cx.open_url(ABOUT_GITHUB_URL);
-}
 
 fn editor_text_font() -> Font {
     // FontFallbacks is internally `Arc<Vec<String>>` — building it once
@@ -149,222 +140,6 @@ fn footnote_row_top_gap(previous: Option<RenderedRowSpacingInfo>, default_gap: f
     } else {
         default_gap
     }
-}
-
-fn is_wide_menu_char(ch: char) -> bool {
-    matches!(
-        ch as u32,
-        0x1100..=0x11ff
-            | 0x2e80..=0xa4cf
-            | 0xac00..=0xd7a3
-            | 0xf900..=0xfaff
-            | 0xfe10..=0xfe6f
-            | 0xff00..=0xff60
-            | 0xffe0..=0xffe6
-    )
-}
-
-fn estimated_menu_label_width(label: &str, text_size: f32) -> f32 {
-    label
-        .chars()
-        .map(|ch| {
-            if ch.is_ascii_whitespace() {
-                text_size * 0.35
-            } else if ch.is_ascii_punctuation() {
-                text_size * 0.45
-            } else if ch.is_ascii() {
-                text_size * 0.62
-            } else if is_wide_menu_char(ch) {
-                text_size
-            } else {
-                text_size * 0.85
-            }
-        })
-        .sum()
-}
-
-fn menu_bar_button_width(label: &str, dimensions: &ThemeDimensions) -> f32 {
-    let content_width = estimated_menu_label_width(label, dimensions.menu_text_size)
-        + dimensions.menu_bar_button_padding_x * 2.0;
-    dimensions.menu_bar_button_width.max(content_width.ceil())
-}
-
-fn supports_in_window_menu_for_target_os(target_os: &str) -> bool {
-    target_os != "macos"
-}
-
-fn supports_in_window_menu() -> bool {
-    supports_in_window_menu_for_target_os(std::env::consts::OS)
-}
-
-fn in_window_menu_bar_height_for_target_os(
-    target_os: &str,
-    has_menus: bool,
-    dimensions: &ThemeDimensions,
-) -> f32 {
-    if has_menus && supports_in_window_menu_for_target_os(target_os) {
-        dimensions.menu_bar_height
-    } else {
-        0.0
-    }
-}
-
-fn menu_panel_left<S: AsRef<str>>(
-    open_index: usize,
-    menu_labels: &[S],
-    dimensions: &ThemeDimensions,
-) -> f32 {
-    let prior_width: f32 = menu_labels
-        .iter()
-        .take(open_index)
-        .map(|label| menu_bar_button_width(label.as_ref(), dimensions))
-        .sum();
-    dimensions.menu_bar_padding_x + prior_width + dimensions.menu_bar_gap * open_index as f32
-}
-
-fn menu_panel_width_for_labels<S: AsRef<str>>(labels: &[S], dimensions: &ThemeDimensions) -> f32 {
-    let widest_label = labels
-        .iter()
-        .map(|label| estimated_menu_label_width(label.as_ref(), dimensions.menu_text_size))
-        .fold(0.0, f32::max);
-    let content_width = widest_label + dimensions.menu_item_padding_x * 2.0;
-    dimensions.menu_panel_width.max(content_width.ceil())
-}
-
-fn owned_menu_item_labels(items: &[OwnedMenuItem]) -> Vec<String> {
-    items
-        .iter()
-        .filter_map(|item| match item {
-            OwnedMenuItem::Action { name, .. } => Some(name.to_string()),
-            OwnedMenuItem::Submenu(menu) => Some(menu.name.to_string()),
-            OwnedMenuItem::SystemMenu(menu) => Some(menu.name.to_string()),
-            OwnedMenuItem::Separator => None,
-        })
-        .collect()
-}
-
-fn menu_item_visual_height(item: &OwnedMenuItem, dimensions: &ThemeDimensions) -> f32 {
-    match item {
-        OwnedMenuItem::Separator => {
-            dimensions.menu_separator_height + dimensions.menu_separator_margin_y * 2.0
-        }
-        OwnedMenuItem::Action { .. } | OwnedMenuItem::Submenu(_) | OwnedMenuItem::SystemMenu(_) => {
-            dimensions.menu_item_height
-        }
-    }
-}
-
-const SCROLLABLE_IMPORT_MENU_VISIBLE_ITEMS: usize = 12;
-
-fn menu_items_visual_height_with_gaps(
-    items: &[OwnedMenuItem],
-    dimensions: &ThemeDimensions,
-) -> f32 {
-    if items.is_empty() {
-        return 0.0;
-    }
-
-    let items_height: f32 = items
-        .iter()
-        .map(|item| menu_item_visual_height(item, dimensions))
-        .sum();
-    items_height + dimensions.menu_panel_gap * items.len().saturating_sub(1) as f32
-}
-
-fn import_menu_split_index(items: &[OwnedMenuItem]) -> Option<usize> {
-    let [
-        prefix @ ..,
-        OwnedMenuItem::Separator,
-        OwnedMenuItem::Action { action, .. },
-    ] = items
-    else {
-        return None;
-    };
-
-    if action.as_ref().as_any().is::<AddThemeConfig>()
-        || action.as_ref().as_any().is::<AddLanguageConfig>()
-    {
-        Some(prefix.len())
-    } else {
-        None
-    }
-}
-
-fn scrollable_import_menu_scroll_height(
-    scroll_items: &[OwnedMenuItem],
-    footer_items: &[OwnedMenuItem],
-    viewport_height: f32,
-    top_offset: f32,
-    dimensions: &ThemeDimensions,
-) -> f32 {
-    let visible_count = scroll_items.len().min(SCROLLABLE_IMPORT_MENU_VISIBLE_ITEMS);
-    if visible_count == 0 {
-        return 0.0;
-    }
-
-    let default_height =
-        menu_items_visual_height_with_gaps(&scroll_items[..visible_count], dimensions);
-    let footer_height = menu_items_visual_height_with_gaps(footer_items, dimensions);
-    let footer_gap = if footer_items.is_empty() {
-        0.0
-    } else {
-        dimensions.menu_panel_gap
-    };
-    let available_height = viewport_height
-        - top_offset
-        - dimensions.menu_panel_top
-        - dimensions.menu_panel_padding * 2.0
-        - footer_height
-        - footer_gap
-        - 8.0;
-    let min_height = dimensions.menu_item_height.min(default_height).max(1.0);
-
-    default_height.min(available_height.max(min_height))
-}
-
-fn submenu_panel_top(
-    items: &[OwnedMenuItem],
-    item_index: usize,
-    dimensions: &ThemeDimensions,
-) -> f32 {
-    let prior_items_height: f32 = items
-        .iter()
-        .take(item_index)
-        .map(|item| menu_item_visual_height(item, dimensions))
-        .sum();
-    let prior_gaps = dimensions.menu_panel_gap * item_index as f32;
-    dimensions.menu_panel_top + dimensions.menu_panel_padding + prior_items_height + prior_gaps
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-struct MenuSubmenuBridgeGeometry {
-    left: f32,
-    top: f32,
-    width: f32,
-    height: f32,
-}
-
-fn submenu_bridge_geometry<S: AsRef<str>, T: AsRef<str>>(
-    open_index: usize,
-    menu_labels: &[S],
-    items: &[OwnedMenuItem],
-    item_index: usize,
-    submenu_labels: &[T],
-    dimensions: &ThemeDimensions,
-) -> Option<MenuSubmenuBridgeGeometry> {
-    let item = items.get(item_index)?;
-    let main_panel_left = menu_panel_left(open_index, menu_labels, dimensions);
-    let main_panel_width = menu_panel_width_for_labels(&owned_menu_item_labels(items), dimensions);
-    let submenu_width = menu_panel_width_for_labels(submenu_labels, dimensions);
-    let vertical_tolerance = dimensions.menu_panel_padding + dimensions.menu_panel_gap;
-    let item_top = submenu_panel_top(items, item_index, dimensions);
-    let top = (item_top - vertical_tolerance).max(dimensions.menu_panel_top);
-    Some(MenuSubmenuBridgeGeometry {
-        left: main_panel_left + main_panel_width,
-        top,
-        width: dimensions.menu_panel_gap + submenu_width,
-        height: menu_item_visual_height(item, dimensions) + vertical_tolerance * 2.0,
-    })
 }
 
 fn footnote_group_shell(
@@ -535,662 +310,6 @@ impl Editor {
             }
         }
     }
-
-    // DEPRECATED
-    // fn sync_window_title(&mut self, window: &mut Window, strings: &I18nStrings) {
-    //     if self.pending_window_title_refresh {
-    //         self.pending_window_title_refresh = false;
-    //         let title = Self::window_title(self.file_path.as_deref(), self.document_dirty, strings);
-    //         window.set_window_title(&title);
-    //     }
-    // }
-
-    /// Renders the in-window fallback menu bar backed by the app menus
-    /// registered through `App::set_menus`. `menus` and `menu_labels` are
-    /// fetched and computed once at the caller and shared with
-    /// [`Self::render_in_window_menu_panel`].
-    fn render_in_window_menu_bar(
-        &self,
-        theme: &Theme,
-        cx: &mut Context<Self>,
-        menus: Option<&[gpui::OwnedMenu]>,
-        menu_labels: &[SharedString],
-        top_offset: f32,
-    ) -> Option<AnyElement> {
-        let menus = menus?;
-        if menus.is_empty() {
-            return None;
-        }
-
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let editor = cx.entity().downgrade();
-        let button_widths = menu_labels
-            .iter()
-            .map(|label| menu_bar_button_width(label, d))
-            .collect::<Vec<_>>();
-
-        Some(
-            div()
-                .id("app-menu-bar")
-                .absolute()
-                .top(px(top_offset))
-                .left_0()
-                .right_0()
-                .h(px(d.menu_bar_height))
-                .occlude()
-                .flex()
-                .items_center()
-                .gap(px(d.menu_bar_gap))
-                .px(px(d.menu_bar_padding_x))
-                .py(px(d.menu_bar_padding_y))
-                .bg(c.dialog_surface)
-                .border_b(px(theme.dimensions.dialog_border_width))
-                .border_color(c.dialog_border)
-                .on_hover(cx.listener(Self::on_menu_bar_hover))
-                .children(menu_labels.iter().enumerate().map(|(index, label)| {
-                    let label = label.clone();
-                    let is_open = self.menu_bar_open == Some(index);
-                    let button_editor = editor.clone();
-                    let button_width = button_widths[index];
-
-                    div()
-                        .id(("app-menu-button", index))
-                        .h(px(d.menu_bar_button_height))
-                        .w(px(button_width))
-                        .px(px(d.menu_bar_button_padding_x))
-                        .flex()
-                        .flex_shrink_0()
-                        .items_center()
-                        .justify_center()
-                        .rounded(px(d.menu_bar_button_radius))
-                        .bg(if is_open {
-                            c.dialog_secondary_button_hover
-                        } else {
-                            c.dialog_surface
-                        })
-                        .hover(|this| this.bg(c.dialog_secondary_button_hover))
-                        .active(|this| this.opacity(0.92))
-                        .cursor_pointer()
-                        .text_size(px(d.menu_text_size))
-                        .font_weight(t.dialog_button_weight.to_font_weight())
-                        .text_color(c.dialog_secondary_button_text)
-                        .whitespace_nowrap()
-                        .child(label)
-                        .on_hover(move |hovered, _window, cx| {
-                            if *hovered {
-                                let _ = button_editor
-                                    .update(cx, |editor, cx| editor.open_menu_bar(index, cx));
-                            }
-                        })
-                }))
-                .into_any_element(),
-        )
-    }
-
-    fn render_in_window_menu_item(
-        &self,
-        item: OwnedMenuItem,
-        item_index: usize,
-        theme: &Theme,
-        editor: WeakEntity<Self>,
-    ) -> AnyElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-
-        match item {
-            OwnedMenuItem::Separator => div()
-                .id(("app-menu-separator", item_index))
-                .flex_shrink_0()
-                .mx(px(d.menu_separator_margin_x))
-                .my(px(d.menu_separator_margin_y))
-                .h(px(d.menu_separator_height))
-                .bg(c.dialog_border)
-                .into_any_element(),
-            OwnedMenuItem::Action { name, action, .. } => {
-                let is_disabled = action.as_ref().as_any().is::<NoRecentFiles>();
-                let click_editor = editor.clone();
-                let hover_editor = editor.clone();
-                let base = div()
-                    .id(("app-menu-item", item_index))
-                    .w_full()
-                    .h(px(d.menu_item_height))
-                    .flex_shrink_0()
-                    .px(px(d.menu_item_padding_x))
-                    .flex()
-                    .items_center()
-                    .rounded(px(d.menu_item_radius))
-                    .bg(c.dialog_surface)
-                    .text_size(px(d.menu_text_size))
-                    .font_weight(t.dialog_body_weight.to_font_weight())
-                    .text_color(if is_disabled {
-                        c.dialog_muted
-                    } else {
-                        c.dialog_secondary_button_text
-                    })
-                    .child(name)
-                    .on_hover(move |hovered, _window, cx| {
-                        if *hovered {
-                            let _ =
-                                hover_editor.update(cx, |editor, cx| editor.close_menu_submenu(cx));
-                        }
-                    });
-
-                if is_disabled {
-                    base.into_any_element()
-                } else {
-                    base.hover(|this| this.bg(c.dialog_secondary_button_hover))
-                        .active(|this| this.opacity(0.92))
-                        .cursor_pointer()
-                        .on_click(move |_, window, cx| {
-                            let _ = click_editor.update(cx, |editor, cx| editor.close_menu_bar(cx));
-                            dispatch_menu_action_for_editor(
-                                action.as_ref(),
-                                &click_editor,
-                                window,
-                                cx,
-                            );
-                        })
-                        .into_any_element()
-                }
-            }
-            OwnedMenuItem::Submenu(submenu) => {
-                let is_open = self.menu_submenu_open == Some(item_index);
-                let hover_editor = editor.clone();
-                div()
-                    .id(("app-menu-submenu", item_index))
-                    .w_full()
-                    .h(px(d.menu_item_height))
-                    .flex_shrink_0()
-                    .px(px(d.menu_item_padding_x))
-                    .flex()
-                    .items_center()
-                    .justify_between()
-                    .rounded(px(d.menu_item_radius))
-                    .bg(if is_open {
-                        c.dialog_secondary_button_hover
-                    } else {
-                        c.dialog_surface
-                    })
-                    .hover(|this| this.bg(c.dialog_secondary_button_hover))
-                    .cursor_pointer()
-                    .text_size(px(d.menu_text_size))
-                    .font_weight(t.dialog_body_weight.to_font_weight())
-                    .text_color(c.dialog_secondary_button_text)
-                    .child(submenu.name.to_string())
-                    .child(">")
-                    .on_hover(move |hovered, _window, cx| {
-                        if *hovered {
-                            let _ = hover_editor
-                                .update(cx, |editor, cx| editor.open_menu_submenu(item_index, cx));
-                        }
-                    })
-                    .into_any_element()
-            }
-            OwnedMenuItem::SystemMenu(os_menu) => div()
-                .id(("app-menu-system", item_index))
-                .w_full()
-                .h(px(d.menu_item_height))
-                .flex_shrink_0()
-                .px(px(d.menu_item_padding_x))
-                .flex()
-                .items_center()
-                .rounded(px(d.menu_item_radius))
-                .bg(c.dialog_surface)
-                .text_size(px(d.menu_text_size))
-                .text_color(c.dialog_muted)
-                .child(os_menu.name.to_string())
-                .into_any_element(),
-        }
-    }
-
-    /// Renders the currently open in-window fallback menu as a floating
-    /// panel. `menus` and `menu_labels` are fetched and computed once at
-    /// the caller and shared with [`Self::render_in_window_menu_bar`].
-    fn render_in_window_menu_panel(
-        &self,
-        theme: &Theme,
-        cx: &mut Context<Self>,
-        menus: Option<&[gpui::OwnedMenu]>,
-        menu_labels: &[SharedString],
-        top_offset: f32,
-        viewport_height: f32,
-    ) -> Option<AnyElement> {
-        let open_index = self.menu_bar_open?;
-        let menus = menus?;
-        let menu = menus.get(open_index)?.clone();
-        let menu_items = menu.items.clone();
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let editor = cx.entity().downgrade();
-        let menu_item_labels = owned_menu_item_labels(&menu_items);
-        let menu_panel_width = menu_panel_width_for_labels(&menu_item_labels, d);
-        let submenu_bridge = self.menu_submenu_open.and_then(|submenu_index| {
-            match menu_items.get(submenu_index)? {
-                OwnedMenuItem::Submenu(submenu) => {
-                    let submenu_labels = owned_menu_item_labels(&submenu.items);
-                    let geometry = submenu_bridge_geometry(
-                        open_index,
-                        menu_labels,
-                        &menu_items,
-                        submenu_index,
-                        &submenu_labels,
-                        d,
-                    )?;
-                    Some(
-                        div()
-                            .id(("app-submenu-bridge", open_index * 1000 + submenu_index))
-                            .absolute()
-                            .occlude()
-                            .top(px(top_offset + geometry.top))
-                            .left(px(geometry.left))
-                            .w(px(geometry.width))
-                            .h(px(geometry.height))
-                            .bg(hsla(0.0, 0.0, 0.0, 0.0))
-                            .on_hover(cx.listener(Self::on_menu_submenu_bridge_hover))
-                            .into_any_element(),
-                    )
-                }
-                _ => None,
-            }
-        });
-        let submenu_panel =
-            self.menu_submenu_open.and_then(|submenu_index| {
-                match menu_items.get(submenu_index)? {
-                    OwnedMenuItem::Submenu(submenu) => {
-                        let submenu_labels = owned_menu_item_labels(&submenu.items);
-                        let left = menu_panel_left(open_index, menu_labels, d)
-                            + menu_panel_width
-                            + d.menu_panel_gap;
-                        let top = submenu_panel_top(&menu_items, submenu_index, d);
-                        let submenu_width = menu_panel_width_for_labels(&submenu_labels, d);
-                        let submenu_items = submenu.items.clone().into_iter().enumerate().map(
-                            |(item_index, item)| match item {
-                                OwnedMenuItem::Separator => div()
-                                    .id((
-                                        "app-submenu-separator",
-                                        submenu_index * 1000 + item_index,
-                                    ))
-                                    .mx(px(d.menu_separator_margin_x))
-                                    .my(px(d.menu_separator_margin_y))
-                                    .h(px(d.menu_separator_height))
-                                    .bg(c.dialog_border)
-                                    .into_any_element(),
-                                OwnedMenuItem::Action { name, action, .. } => {
-                                    let is_disabled =
-                                        action.as_ref().as_any().is::<NoRecentFiles>();
-                                    let editor = editor.clone();
-                                    let base = div()
-                                        .id(("app-submenu-item", submenu_index * 1000 + item_index))
-                                        .w_full()
-                                        .h(px(d.menu_item_height))
-                                        .px(px(d.menu_item_padding_x))
-                                        .flex()
-                                        .items_center()
-                                        .rounded(px(d.menu_item_radius))
-                                        .bg(c.dialog_surface)
-                                        .text_size(px(d.menu_text_size))
-                                        .font_weight(t.dialog_body_weight.to_font_weight())
-                                        .text_color(if is_disabled {
-                                            c.dialog_muted
-                                        } else {
-                                            c.dialog_secondary_button_text
-                                        })
-                                        .child(name);
-
-                                    if is_disabled {
-                                        base.into_any_element()
-                                    } else {
-                                        base.hover(|this| this.bg(c.dialog_secondary_button_hover))
-                                            .active(|this| this.opacity(0.92))
-                                            .cursor_pointer()
-                                            .on_click(move |_, window, cx| {
-                                                let _ = editor.update(cx, |editor, cx| {
-                                                    editor.close_menu_bar(cx)
-                                                });
-                                                dispatch_menu_action_for_editor(
-                                                    action.as_ref(),
-                                                    &editor,
-                                                    window,
-                                                    cx,
-                                                );
-                                            })
-                                            .into_any_element()
-                                    }
-                                }
-                                OwnedMenuItem::Submenu(submenu) => div()
-                                    .id(("app-submenu-nested", submenu_index * 1000 + item_index))
-                                    .w_full()
-                                    .h(px(d.menu_item_height))
-                                    .px(px(d.menu_item_padding_x))
-                                    .flex()
-                                    .items_center()
-                                    .rounded(px(d.menu_item_radius))
-                                    .bg(c.dialog_surface)
-                                    .text_size(px(d.menu_text_size))
-                                    .text_color(c.dialog_muted)
-                                    .child(submenu.name.to_string())
-                                    .into_any_element(),
-                                OwnedMenuItem::SystemMenu(os_menu) => div()
-                                    .id(("app-submenu-system", submenu_index * 1000 + item_index))
-                                    .w_full()
-                                    .h(px(d.menu_item_height))
-                                    .px(px(d.menu_item_padding_x))
-                                    .flex()
-                                    .items_center()
-                                    .rounded(px(d.menu_item_radius))
-                                    .bg(c.dialog_surface)
-                                    .text_size(px(d.menu_text_size))
-                                    .text_color(c.dialog_muted)
-                                    .child(os_menu.name.to_string())
-                                    .into_any_element(),
-                            },
-                        );
-
-                        Some(
-                            div()
-                                .id(("app-submenu-panel", open_index * 1000 + submenu_index))
-                                .absolute()
-                                .occlude()
-                                .top(px(top_offset + top))
-                                .left(px(left))
-                                .w(px(submenu_width))
-                                .p(px(d.menu_panel_padding))
-                                .flex()
-                                .flex_col()
-                                .gap(px(d.menu_panel_gap))
-                                .bg(c.dialog_surface)
-                                .border(px(d.dialog_border_width))
-                                .border_color(c.dialog_border)
-                                .rounded(px(d.menu_panel_radius))
-                                .shadow_lg()
-                                .on_hover(cx.listener(Self::on_menu_submenu_panel_hover))
-                                .children(submenu_items)
-                                .into_any_element(),
-                        )
-                    }
-                    _ => None,
-                }
-            });
-
-        let main_panel = div()
-            .id(("app-menu-panel", open_index))
-            .absolute()
-            .occlude()
-            .top(px(top_offset + d.menu_panel_top))
-            .left(px(menu_panel_left(open_index, menu_labels, d)))
-            .w(px(menu_panel_width))
-            .p(px(d.menu_panel_padding))
-            .flex()
-            .flex_col()
-            .gap(px(d.menu_panel_gap))
-            .bg(c.dialog_surface)
-            .border(px(d.dialog_border_width))
-            .border_color(c.dialog_border)
-            .rounded(px(d.menu_panel_radius))
-            .shadow_lg()
-            .on_hover(cx.listener(Self::on_menu_panel_hover));
-        let main_panel = if let Some(split_index) = import_menu_split_index(&menu_items) {
-            let scroll_items = &menu_items[..split_index];
-            let footer_items = &menu_items[split_index..];
-            let scroll_height = scrollable_import_menu_scroll_height(
-                scroll_items,
-                footer_items,
-                viewport_height,
-                top_offset,
-                d,
-            );
-            let scroll_area = (!scroll_items.is_empty()).then(|| {
-                div()
-                    .id(("app-menu-scroll-area", open_index))
-                    .w_full()
-                    .h(px(scroll_height))
-                    .flex_shrink_0()
-                    .min_h(px(0.0))
-                    .overflow_y_scroll()
-                    .child(
-                        div()
-                            .w_full()
-                            .flex()
-                            .flex_col()
-                            .gap(px(d.menu_panel_gap))
-                            .children(scroll_items.iter().cloned().enumerate().map(
-                                |(item_index, item)| {
-                                    self.render_in_window_menu_item(
-                                        item,
-                                        item_index,
-                                        theme,
-                                        editor.clone(),
-                                    )
-                                },
-                            )),
-                    )
-                    .into_any_element()
-            });
-            let footer_elements =
-                footer_items
-                    .iter()
-                    .cloned()
-                    .enumerate()
-                    .map(|(footer_index, item)| {
-                        self.render_in_window_menu_item(
-                            item,
-                            split_index + footer_index,
-                            theme,
-                            editor.clone(),
-                        )
-                    });
-
-            main_panel
-                .children(scroll_area)
-                .children(footer_elements)
-                .into_any_element()
-        } else {
-            let items = menu_items
-                .iter()
-                .cloned()
-                .enumerate()
-                .map(|(item_index, item)| {
-                    self.render_in_window_menu_item(item, item_index, theme, editor.clone())
-                });
-
-            main_panel.children(items).into_any_element()
-        };
-
-        let layer = div()
-            .id(("app-menu-panel-layer", open_index))
-            .absolute()
-            .top_0()
-            .left_0()
-            .right_0()
-            .bottom_0()
-            .child(main_panel);
-        let layer = if let Some(submenu_bridge) = submenu_bridge {
-            layer.child(submenu_bridge)
-        } else {
-            layer
-        };
-        let layer = if let Some(submenu_panel) = submenu_panel {
-            layer.child(submenu_panel)
-        } else {
-            layer
-        };
-
-        Some(layer.into_any_element())
-    }
-
-    fn info_dialog_title<'a>(&self, strings: &'a I18nStrings, kind: InfoDialogKind) -> &'a str {
-        match kind {
-            InfoDialogKind::CheckForUpdates => &strings.help_check_updates_title,
-            InfoDialogKind::About => &strings.help_about_title,
-        }
-    }
-
-    pub(crate) fn about_dialog_body_lines(strings: &I18nStrings) -> Vec<String> {
-        vec![
-            format!("Velotype {}", env!("CARGO_PKG_VERSION")),
-            strings.help_about_message.clone(),
-            format!("{}: {}", strings.help_about_github_label, ABOUT_GITHUB_URL),
-            strings.help_about_star_message.clone(),
-        ]
-    }
-
-    fn info_dialog_body(&self, strings: &I18nStrings, kind: InfoDialogKind) -> String {
-        match kind {
-            InfoDialogKind::CheckForUpdates => strings.help_check_updates_message.clone(),
-            InfoDialogKind::About => Self::about_dialog_body_lines(strings).join("\n"),
-        }
-    }
-
-    fn render_info_dialog_body(
-        &self,
-        theme: &Theme,
-        strings: &I18nStrings,
-        kind: InfoDialogKind,
-    ) -> AnyElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let body_style = |this: Div| {
-            this.text_size(px(t.dialog_body_size))
-                .font_weight(t.dialog_body_weight.to_font_weight())
-                .line_height(rems(t.text_line_height))
-                .text_color(c.dialog_body)
-        };
-
-        match kind {
-            InfoDialogKind::CheckForUpdates => div()
-                .flex()
-                .flex_col()
-                .gap(px(d.dialog_gap * 0.5))
-                .child(
-                    body_style(div()).children(
-                        self.info_dialog_body(strings, kind)
-                            .lines()
-                            .map(|line| div().child(line.to_string())),
-                    ),
-                )
-                .into_any_element(),
-            InfoDialogKind::About => div()
-                .flex()
-                .flex_col()
-                .gap(px(d.dialog_gap * 0.5))
-                .child(body_style(div()).child(format!("Velotype {}", env!("CARGO_PKG_VERSION"))))
-                .child(body_style(div()).child(strings.help_about_message.clone()))
-                .child(
-                    body_style(div())
-                        .flex()
-                        .flex_wrap()
-                        .gap(px(4.0))
-                        .child(format!("{}:", strings.help_about_github_label))
-                        .child(
-                            div()
-                                .id("about-github-link")
-                                .cursor_pointer()
-                                .text_color(c.text_link)
-                                .underline()
-                                .child(ABOUT_GITHUB_URL)
-                                .on_click(move |_, _, cx| {
-                                    open_about_github_url(cx);
-                                }),
-                        ),
-                )
-                .child(body_style(div()).child(strings.help_about_star_message.clone()))
-                .into_any_element(),
-        }
-    }
-
-    fn on_dismiss_info_dialog(
-        &mut self,
-        _: &ClickEvent,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.hide_info_dialog(cx);
-    }
-
-    fn render_info_dialog_overlay(
-        &self,
-        theme: &Theme,
-        kind: InfoDialogKind,
-        cx: &mut Context<Self>,
-    ) -> impl IntoElement {
-        let c = &theme.colors;
-        let d = &theme.dimensions;
-        let t = &theme.typography;
-        let strings = cx.global::<I18nManager>().strings();
-
-        div()
-            .id("info-dialog-overlay")
-            .absolute()
-            .top_0()
-            .left_0()
-            .right_0()
-            .bottom_0()
-            .occlude()
-            .flex()
-            .items_center()
-            .justify_center()
-            .bg(c.dialog_backdrop)
-            .child(
-                div()
-                    .w_full()
-                    .px(px(d.editor_padding))
-                    .flex()
-                    .justify_center()
-                    .child(
-                        div()
-                            .id("info-dialog")
-                            .w(px(d.dialog_width))
-                            .max_w(relative(1.0))
-                            .flex()
-                            .flex_col()
-                            .gap(px(d.dialog_gap))
-                            .p(px(d.dialog_padding))
-                            .bg(c.dialog_surface)
-                            .border(px(d.dialog_border_width))
-                            .border_color(c.dialog_border)
-                            .rounded(px(d.dialog_radius))
-                            .shadow_lg()
-                            .child(
-                                div()
-                                    .text_size(px(t.dialog_title_size))
-                                    .font_weight(t.dialog_title_weight.to_font_weight())
-                                    .text_color(c.dialog_title)
-                                    .child(self.info_dialog_title(strings, kind).to_string()),
-                            )
-                            .child(self.render_info_dialog_body(theme, strings, kind))
-                            .child(
-                                div()
-                                    .flex()
-                                    .justify_end()
-                                    .gap(px(d.dialog_button_gap))
-                                    .child(
-                                        div()
-                                            .id("dismiss-info-dialog")
-                                            .h(px(d.dialog_button_height))
-                                            .px(px(d.dialog_button_padding_x))
-                                            .flex()
-                                            .items_center()
-                                            .justify_center()
-                                            .rounded(px((d.dialog_radius - 4.0).max(0.0)))
-                                            .bg(c.dialog_primary_button_bg)
-                                            .hover(|this| this.bg(c.dialog_primary_button_hover))
-                                            .active(|this| this.opacity(0.92))
-                                            .cursor_pointer()
-                                            .text_size(px(t.dialog_button_size))
-                                            .font_weight(t.dialog_button_weight.to_font_weight())
-                                            .text_color(c.dialog_primary_button_text)
-                                            .child(strings.info_dialog_ok.clone())
-                                            .on_click(cx.listener(Self::on_dismiss_info_dialog)),
-                                    ),
-                            ),
-                    ),
-            )
-    }
 }
 
 impl Render for Editor {
@@ -1215,13 +334,6 @@ impl Render for Editor {
         let d = &theme.dimensions;
         let visible_blocks = self.document.visible_blocks().to_vec();
         let editor = cx.entity().downgrade();
-        let has_menus = cx
-            .get_menus()
-            .map(|menus| !menus.is_empty())
-            .unwrap_or(false);
-        let titlebar_height = custom_titlebar_height(window, d);
-        let menu_bar_height =
-            in_window_menu_bar_height_for_target_os(std::env::consts::OS, has_menus, d);
         let scroll_trigger_padding = (d.block_min_height * 0.75).max(16.0);
         let max_scroll_y = f32::from(self.scroll_handle.max_offset().height.max(px(0.0)));
         let viewport_height = f32::from(viewport_bounds.size.height.max(px(1.0)));
@@ -1695,39 +807,9 @@ impl Render for Editor {
             .on_action(cx.listener(Self::on_page_down))
             .on_action(cx.listener(Self::on_jump_to_top))
             .on_action(cx.listener(Self::on_jump_to_bottom));
-        // Fetch menus + collect labels once for both renderers; previously each
-        // of render_in_window_menu_bar / render_in_window_menu_panel called
-        // cx.get_menus() and walked menus.iter().map(|m| m.name.to_string())
-        // independently — two redundant Vec<OwnedMenu> + two redundant
-        // Vec<String>-of-N-allocations per frame.
-        let menus = supports_in_window_menu()
-            .then(|| cx.get_menus())
-            .flatten()
-            .filter(|m| !m.is_empty());
-        let menu_labels: Vec<SharedString> = menus
-            .as_ref()
-            .map(|m| m.iter().map(|menu| menu.name.clone()).collect())
-            .unwrap_or_default();
-        let base = if let Some(menu_bar) = self.render_in_window_menu_bar(
-            &theme,
-            cx,
-            menus.as_deref(),
-            &menu_labels,
-            titlebar_height,
-        ) {
-            base.child(menu_bar)
-        } else {
-            base
-        };
         let workspace_width =
             workspace_panel_width_for_viewport(f32::from(window.viewport_size().width));
-        let main_content = div()
-            .w_full()
-            .flex_1()
-            .min_h(px(0.0))
-            .pt(px(titlebar_height + menu_bar_height))
-            .flex()
-            .min_w(px(0.0));
+        let main_content = div().w_full().flex_1().min_h(px(0.0)).flex().min_w(px(0.0));
         let main_content = if let Some(workspace_panel) =
             self.render_workspace_panel(&theme, &strings, workspace_width, cx)
         {
@@ -1736,433 +818,7 @@ impl Render for Editor {
             main_content
         };
         let base = base.child(main_content.child(content_area));
-        let base = if let Some(menu_panel) = self.render_in_window_menu_panel(
-            &theme,
-            cx,
-            menus.as_deref(),
-            &menu_labels,
-            titlebar_height,
-            f32::from(window.viewport_size().height.max(px(1.0))),
-        ) {
-            base.child(menu_panel)
-        } else {
-            base
-        };
-        if let Some(kind) = self.info_dialog {
-            base.child(self.render_info_dialog_overlay(&theme, kind, cx))
-        } else {
-            base
-        }
-    }
-}
 
-#[cfg(test)]
-mod tests {
-    use super::{
-        NoRecentFiles, RenderedRowSpacingInfo, callout_row_top_gap, editor_text_font,
-        import_menu_split_index, in_window_menu_bar_height_for_target_os, menu_bar_button_width,
-        menu_items_visual_height_with_gaps, menu_panel_left, menu_panel_width_for_labels,
-        owned_menu_item_labels, rendered_row_top_gap, scrollable_import_menu_scroll_height,
-        submenu_bridge_geometry, supports_in_window_menu_for_target_os,
-        tibetan_font_fallbacks_for_target_os,
-    };
-    use crate::components::{AddLanguageConfig, AddThemeConfig};
-    use crate::theme::Theme;
-    use gpui::{OwnedMenu, OwnedMenuItem};
-    use uuid::Uuid;
-
-    fn disabled_menu_action(name: &str) -> OwnedMenuItem {
-        OwnedMenuItem::Action {
-            name: name.into(),
-            action: Box::new(NoRecentFiles),
-            os_action: None,
-        }
-    }
-
-    fn add_theme_menu_action() -> OwnedMenuItem {
-        OwnedMenuItem::Action {
-            name: "Add Theme Config".into(),
-            action: Box::new(AddThemeConfig),
-            os_action: None,
-        }
-    }
-
-    fn add_language_menu_action() -> OwnedMenuItem {
-        OwnedMenuItem::Action {
-            name: "Add Language Config".into(),
-            action: Box::new(AddLanguageConfig),
-            os_action: None,
-        }
-    }
-
-    #[test]
-    fn contiguous_quote_rows_collapse_inter_row_gap() {
-        let group = Uuid::new_v4();
-        let gap = rendered_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                quote_group_anchor: Some(group),
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo {
-                quote_group_anchor: Some(group),
-                ..RenderedRowSpacingInfo::default()
-            },
-            4.0,
-        );
-        assert_eq!(gap, 0.0);
-    }
-
-    #[test]
-    fn editor_text_font_keeps_system_ui_as_primary_family() {
-        assert_eq!(editor_text_font().family.to_string(), ".SystemUIFont");
-    }
-
-    #[test]
-    fn tibetan_font_fallbacks_prioritize_platform_defaults() {
-        assert_eq!(
-            tibetan_font_fallbacks_for_target_os("windows")
-                .first()
-                .map(String::as_str),
-            Some("Microsoft Himalaya")
-        );
-        assert_eq!(
-            tibetan_font_fallbacks_for_target_os("macos")
-                .first()
-                .map(String::as_str),
-            Some("Kailasa")
-        );
-        assert_eq!(
-            tibetan_font_fallbacks_for_target_os("linux")
-                .first()
-                .map(String::as_str),
-            Some("Noto Serif Tibetan")
-        );
-        assert_eq!(
-            tibetan_font_fallbacks_for_target_os("unknown")
-                .first()
-                .map(String::as_str),
-            Some("Noto Serif Tibetan")
-        );
-    }
-
-    #[test]
-    fn nested_quote_separator_row_keeps_outer_group_gap_collapsed() {
-        let group = Uuid::new_v4();
-        let gap = rendered_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                quote_group_anchor: Some(group),
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo {
-                quote_group_anchor: Some(group),
-                ..RenderedRowSpacingInfo::default()
-            },
-            4.0,
-        );
-        assert_eq!(gap, 0.0);
-    }
-
-    #[test]
-    fn distinct_quote_groups_keep_default_gap() {
-        let gap = rendered_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                quote_group_anchor: Some(Uuid::new_v4()),
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo {
-                quote_group_anchor: Some(Uuid::new_v4()),
-                ..RenderedRowSpacingInfo::default()
-            },
-            4.0,
-        );
-        assert_eq!(gap, 4.0);
-    }
-
-    #[test]
-    fn non_quote_rows_keep_default_gap() {
-        let gap = rendered_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                quote_group_anchor: None,
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo {
-                quote_group_anchor: Some(Uuid::new_v4()),
-                ..RenderedRowSpacingInfo::default()
-            },
-            4.0,
-        );
-        assert_eq!(gap, 4.0);
-    }
-
-    #[test]
-    fn callout_inner_spacing_uses_header_and_body_tokens() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-
-        let header_gap = callout_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                is_callout_header: true,
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo::default(),
-            dimensions,
-        );
-        let body_gap = callout_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                is_callout_header: false,
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo::default(),
-            dimensions,
-        );
-
-        assert_eq!(header_gap, dimensions.callout_header_margin_bottom);
-        assert_eq!(body_gap, dimensions.callout_body_gap);
-    }
-
-    #[test]
-    fn nested_quote_rows_inside_callout_collapse_body_gap() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-        let group = Uuid::new_v4();
-
-        let gap = callout_row_top_gap(
-            Some(RenderedRowSpacingInfo {
-                is_callout_header: false,
-                visible_quote_group_anchor: Some(group),
-                ..RenderedRowSpacingInfo::default()
-            }),
-            RenderedRowSpacingInfo {
-                visible_quote_group_anchor: Some(group),
-                ..RenderedRowSpacingInfo::default()
-            },
-            dimensions,
-        );
-
-        assert_eq!(gap, 0.0);
-    }
-
-    #[test]
-    fn menu_button_width_expands_for_long_ascii_labels() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-
-        assert_eq!(
-            menu_bar_button_width("文件", dimensions),
-            dimensions.menu_bar_button_width
-        );
-        assert!(menu_bar_button_width("Language", dimensions) > dimensions.menu_bar_button_width);
-    }
-
-    #[test]
-    fn in_window_menu_is_enabled_for_every_target_except_macos() {
-        for target_os in [
-            "windows",
-            "linux",
-            "freebsd",
-            "openbsd",
-            "netbsd",
-            "dragonfly",
-            "solaris",
-            "illumos",
-            "android",
-            "unknown",
-        ] {
-            assert!(
-                supports_in_window_menu_for_target_os(target_os),
-                "{target_os} should use the in-window fallback menu"
-            );
-        }
-        assert!(!supports_in_window_menu_for_target_os("macos"));
-    }
-
-    #[test]
-    fn in_window_menu_height_depends_on_platform_and_menu_presence() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-
-        assert_eq!(
-            in_window_menu_bar_height_for_target_os("linux", true, dimensions),
-            dimensions.menu_bar_height
-        );
-        assert_eq!(
-            in_window_menu_bar_height_for_target_os("windows", true, dimensions),
-            dimensions.menu_bar_height
-        );
-        assert_eq!(
-            in_window_menu_bar_height_for_target_os("linux", false, dimensions),
-            0.0
-        );
-        assert_eq!(
-            in_window_menu_bar_height_for_target_os("macos", true, dimensions),
-            0.0
-        );
-    }
-
-    #[test]
-    fn menu_panel_left_uses_accumulated_dynamic_button_widths() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-        let labels = vec![
-            "File".to_string(),
-            "Language".to_string(),
-            "Theme".to_string(),
-            "Help".to_string(),
-        ];
-
-        let left = menu_panel_left(2, &labels, dimensions);
-        let expected = dimensions.menu_bar_padding_x
-            + menu_bar_button_width("File", dimensions)
-            + dimensions.menu_bar_gap
-            + menu_bar_button_width("Language", dimensions)
-            + dimensions.menu_bar_gap;
-        let old_fixed_left = dimensions.menu_bar_padding_x
-            + 2.0 * (dimensions.menu_bar_button_width + dimensions.menu_bar_gap);
-
-        assert_eq!(left, expected);
-        assert!(left > old_fixed_left);
-    }
-
-    #[test]
-    fn menu_panel_width_expands_for_long_recent_paths() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-        let short_labels = vec!["Save".to_string()];
-        let long_labels = vec![r"C:\Users\someone\Documents\Very Long Folder\notes.md".to_string()];
-
-        assert_eq!(
-            menu_panel_width_for_labels(&short_labels, dimensions),
-            dimensions.menu_panel_width
-        );
-        assert!(
-            menu_panel_width_for_labels(&long_labels, dimensions) > dimensions.menu_panel_width
-        );
-    }
-
-    #[test]
-    fn import_menu_split_detects_theme_and_language_import_tails() {
-        let theme_items = vec![
-            disabled_menu_action("Velotype"),
-            OwnedMenuItem::Separator,
-            add_theme_menu_action(),
-        ];
-        let language_items = vec![
-            disabled_menu_action("English"),
-            OwnedMenuItem::Separator,
-            add_language_menu_action(),
-        ];
-        let regular_items = vec![
-            disabled_menu_action("Open"),
-            OwnedMenuItem::Separator,
-            disabled_menu_action("Save"),
-        ];
-        let malformed_import_items =
-            vec![disabled_menu_action("Velotype"), add_theme_menu_action()];
-
-        assert_eq!(import_menu_split_index(&theme_items), Some(1));
-        assert_eq!(import_menu_split_index(&language_items), Some(1));
-        assert_eq!(import_menu_split_index(&regular_items), None);
-        assert_eq!(import_menu_split_index(&malformed_import_items), None);
-    }
-
-    #[test]
-    fn scrollable_import_menu_height_caps_visible_items_and_clamps_to_viewport() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-        let scroll_items = (0..20)
-            .map(|index| disabled_menu_action(&format!("Custom Theme {index}")))
-            .collect::<Vec<_>>();
-        let footer_items = vec![OwnedMenuItem::Separator, add_theme_menu_action()];
-        let expected_large_height =
-            menu_items_visual_height_with_gaps(&scroll_items[..12], dimensions);
-        let full_scroll_content_height =
-            menu_items_visual_height_with_gaps(&scroll_items, dimensions);
-        let footer_height = menu_items_visual_height_with_gaps(&footer_items, dimensions);
-
-        let large_height = scrollable_import_menu_scroll_height(
-            &scroll_items,
-            &footer_items,
-            2000.0,
-            0.0,
-            dimensions,
-        );
-        let small_height = scrollable_import_menu_scroll_height(
-            &scroll_items,
-            &footer_items,
-            180.0,
-            0.0,
-            dimensions,
-        );
-
-        assert!((large_height - expected_large_height).abs() < f32::EPSILON);
-        assert!(full_scroll_content_height > large_height);
-        assert!(large_height < expected_large_height + footer_height);
-        assert!(small_height < large_height);
-        assert!(small_height >= dimensions.menu_item_height);
-    }
-
-    #[test]
-    fn submenu_bridge_spans_parent_child_menu_gap() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-        let labels = vec!["File".to_string()];
-        let items = vec![
-            OwnedMenuItem::Separator,
-            OwnedMenuItem::Submenu(OwnedMenu {
-                name: "Recent".into(),
-                items: vec![OwnedMenuItem::Action {
-                    name: r"C:\Users\someone\Documents\notes.md".into(),
-                    action: Box::new(NoRecentFiles),
-                    os_action: None,
-                }],
-            }),
-        ];
-        let submenu_labels = match &items[1] {
-            OwnedMenuItem::Submenu(submenu) => owned_menu_item_labels(&submenu.items),
-            _ => Vec::new(),
-        };
-
-        let bridge = submenu_bridge_geometry(0, &labels, &items, 1, &submenu_labels, dimensions)
-            .expect("submenu bridge geometry should be available");
-        let submenu_width = menu_panel_width_for_labels(&submenu_labels, dimensions);
-
-        assert_eq!(
-            bridge.left,
-            dimensions.menu_bar_padding_x + dimensions.menu_panel_width
-        );
-        assert_eq!(bridge.width, dimensions.menu_panel_gap + submenu_width);
-        assert!(bridge.height > dimensions.menu_item_height);
-        let item_top = dimensions.menu_panel_top
-            + dimensions.menu_panel_padding
-            + dimensions.menu_separator_height
-            + dimensions.menu_separator_margin_y * 2.0
-            + dimensions.menu_panel_gap;
-        assert!(bridge.top < item_top);
-        assert!(bridge.top >= dimensions.menu_panel_top);
-    }
-
-    #[test]
-    fn submenu_bridge_uses_dynamic_main_menu_width() {
-        let theme = Theme::default_theme();
-        let dimensions = &theme.dimensions;
-        let labels = vec!["File".to_string()];
-        let items = vec![OwnedMenuItem::Submenu(OwnedMenu {
-            name: "Open Recently Used Markdown File".into(),
-            items: vec![OwnedMenuItem::Action {
-                name: r"C:\Users\someone\Documents\Very Long Folder\notes.md".into(),
-                action: Box::new(NoRecentFiles),
-                os_action: None,
-            }],
-        })];
-        let submenu_labels = match &items[0] {
-            OwnedMenuItem::Submenu(submenu) => owned_menu_item_labels(&submenu.items),
-            _ => Vec::new(),
-        };
-
-        let bridge = submenu_bridge_geometry(0, &labels, &items, 0, &submenu_labels, dimensions)
-            .expect("submenu bridge geometry should be available");
-
-        assert!(bridge.left > dimensions.menu_bar_padding_x + dimensions.menu_panel_width);
-        assert!(bridge.width > dimensions.menu_panel_gap + dimensions.menu_panel_width);
+        base
     }
 }
