--- crates/opennote-velotype/src/editor/update.rs
@@ -1,158 +1,0 @@
-//! Editor-facing update-check flow.
-
-use futures::FutureExt;
-use futures::channel::oneshot;
-use gpui::*;
-
-use super::{Editor, InfoDialogKind};
-use crate::i18n::I18nManager;
-use crate::net::update::{self as update_check, UpdateCheckResult, UpdateVersionInfo};
-use crate::theme::{BUILTIN_THEME_VELOTYPE_ID, BUILTIN_THEME_VELOTYPE_LIGHT_ID, ThemeManager};
-
-impl Editor {
-    /// Switch the theme between dark and light mode
-    pub fn switch_theme(cx: &mut App, switch_to_dark_mode: bool) {
-        let theme_id = match switch_to_dark_mode {
-            true => BUILTIN_THEME_VELOTYPE_ID,
-            false => BUILTIN_THEME_VELOTYPE_LIGHT_ID,
-        };
-
-        let _ = cx.update_global::<ThemeManager, _>(|theme_manager, _cx| {
-            theme_manager.set_theme_by_id(theme_id);
-        });
-    }
-
-    pub(crate) fn request_check_updates(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        if self.update_check_in_progress {
-            self.show_info_dialog(InfoDialogKind::CheckForUpdates, cx);
-            return;
-        }
-
-        self.update_check_in_progress = true;
-        self.show_info_dialog(InfoDialogKind::CheckForUpdates, cx);
-
-        let weak_editor = cx.entity().downgrade();
-        let window_handle = window.window_handle();
-        let (tx, rx) = oneshot::channel();
-        std::thread::spawn(move || {
-            let result = update_check::check_latest_version(env!("CARGO_PKG_VERSION"));
-            let _ = tx.send(result);
-        });
-
-        cx.spawn(async move |_this: WeakEntity<Self>, cx: &mut AsyncApp| {
-            let result = rx
-                .map(|result| {
-                    result.unwrap_or_else(|_| {
-                        Err(update_check::UpdateCheckError::ParseVersion(
-                            "update check worker ended before returning a result".to_string(),
-                        ))
-                    })
-                })
-                .await;
-
-            let _ = weak_editor.update(cx, |editor, cx| {
-                editor.update_check_in_progress = false;
-                editor.hide_info_dialog(cx);
-            });
-
-            let _ = cx.update_window(
-                window_handle,
-                move |_view: AnyView, window: &mut Window, cx: &mut App| match result {
-                    Ok(UpdateCheckResult::UpdateAvailable(info)) => {
-                        show_update_available_prompt(window, cx, &info);
-                    }
-                    Ok(UpdateCheckResult::UpToDate(info)) => {
-                        show_up_to_date_prompt(window, cx, &info);
-                    }
-                    Err(error) => {
-                        show_update_failed_prompt(window, cx, &error.to_string());
-                    }
-                },
-            );
-        })
-        .detach();
-    }
-}
-
-fn show_update_available_prompt(window: &mut Window, cx: &mut App, info: &UpdateVersionInfo) {
-    let strings = cx.global::<I18nManager>().strings().clone();
-    let detail = format_update_message(
-        &strings.update_available_message_template,
-        &info.current_version,
-        &info.latest_version,
-    );
-    let buttons = [
-        strings.update_open_release.as_str(),
-        strings.update_later.as_str(),
-    ];
-    let prompt = window.prompt(
-        PromptLevel::Info,
-        &strings.update_available_title,
-        Some(&detail),
-        &buttons,
-        cx,
-    );
-    let window_handle = window.window_handle();
-    cx.spawn(async move |cx| {
-        let Ok(choice) = prompt.await else {
-            return;
-        };
-        if choice == 0 {
-            let _ = cx.update_window(window_handle, |_view: AnyView, _window, cx| {
-                cx.open_url(update_check::RELEASES_URL);
-            });
-        }
-    })
-    .detach();
-}
-
-fn show_up_to_date_prompt(window: &mut Window, cx: &mut App, info: &UpdateVersionInfo) {
-    let strings = cx.global::<I18nManager>().strings().clone();
-    let detail = format_update_message(
-        &strings.update_up_to_date_message_template,
-        &info.current_version,
-        &info.latest_version,
-    );
-    let buttons = [strings.info_dialog_ok.as_str()];
-    let _ = window.prompt(
-        PromptLevel::Info,
-        &strings.update_up_to_date_title,
-        Some(&detail),
-        &buttons,
-        cx,
-    );
-}
-
-fn show_update_failed_prompt(window: &mut Window, cx: &mut App, detail: &str) {
-    let strings = cx.global::<I18nManager>().strings().clone();
-    let message = strings
-        .update_failed_message_template
-        .replace("{error}", detail);
-    let buttons = [strings.info_dialog_ok.as_str()];
-    let _ = window.prompt(
-        PromptLevel::Critical,
-        &strings.update_failed_title,
-        Some(&message),
-        &buttons,
-        cx,
-    );
-}
-
-fn format_update_message(template: &str, current_version: &str, latest_version: &str) -> String {
-    template
-        .replace("{current}", current_version)
-        .replace("{latest}", latest_version)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::format_update_message;
-
-    #[test]
-    fn update_message_templates_replace_versions() {
-        assert_eq!(
-            format_update_message("Current {current}, latest {latest}.", "0.2.1", "0.2.2"),
-            "Current 0.2.1, latest 0.2.2."
-        );
-    }
-}
