+++ crates/opennote-velotype/src/editor/window_state.rs
@@ -313,151 +313,11 @@ impl Editor {
             || (f32::from(previous.height) - f32::from(current.height)).abs() > EPSILON
     }
 
-    pub(crate) fn show_info_dialog(&mut self, kind: InfoDialogKind, cx: &mut Context<Self>) {
-        self.menu_bar_open = None;
-        self.menu_submenu_open = None;
-        self.menu_submenu_panel_hovered = false;
-        self.menu_submenu_bridge_hovered = false;
-        self.info_dialog = Some(kind);
-        cx.notify();
-    }
-
-    pub(crate) fn hide_info_dialog(&mut self, cx: &mut Context<Self>) {
-        if self.info_dialog.take().is_some() {
-            cx.notify();
-        }
-    }
-
-    pub(crate) fn open_menu_bar(&mut self, index: usize, cx: &mut Context<Self>) {
-        self.menu_close_task = None;
-        if self.menu_bar_open != Some(index) {
-            self.menu_bar_open = Some(index);
-            self.menu_submenu_open = None;
-            self.menu_submenu_panel_hovered = false;
-            self.menu_submenu_bridge_hovered = false;
-            cx.notify();
-        }
-    }
-
-    pub(crate) fn open_menu_submenu(&mut self, index: usize, cx: &mut Context<Self>) {
-        self.menu_close_task = None;
-        if self.menu_submenu_open != Some(index) {
-            self.menu_submenu_open = Some(index);
-            cx.notify();
-        }
-    }
-
-    pub(crate) fn close_menu_submenu(&mut self, cx: &mut Context<Self>) {
-        let had_open_submenu = self.menu_submenu_open.take().is_some();
-        let had_submenu_hover = self.menu_submenu_panel_hovered || self.menu_submenu_bridge_hovered;
-        self.menu_submenu_panel_hovered = false;
-        self.menu_submenu_bridge_hovered = false;
-        if had_open_submenu || had_submenu_hover {
-            cx.notify();
-        }
-    }
-
-    pub(super) fn schedule_menu_bar_close(&mut self, cx: &mut Context<Self>) {
-        if self.menu_bar_open.is_none() {
-            return;
-        }
-
-        let weak_editor = cx.entity().downgrade();
-        self.menu_close_task = Some(cx.spawn(
-            async move |_this: WeakEntity<Self>, cx: &mut AsyncApp| {
-                cx.background_executor()
-                    .timer(Duration::from_millis(120))
-                    .await;
-                let _ = weak_editor.update(cx, |editor, cx| {
-                    editor.menu_close_task = None;
-                    if !editor.menu_bar_hovered
-                        && !editor.menu_panel_hovered
-                        && !editor.menu_submenu_panel_hovered
-                        && !editor.menu_submenu_bridge_hovered
-                    {
-                        editor.close_menu_bar(cx);
-                    }
-                });
-            },
-        ));
-    }
-
-    pub(crate) fn set_menu_bar_hovered(&mut self, hovered: bool, cx: &mut Context<Self>) {
-        self.menu_bar_hovered = hovered;
-        if hovered {
-            self.menu_close_task = None;
-        } else if !self.menu_panel_hovered
-            && !self.menu_submenu_panel_hovered
-            && !self.menu_submenu_bridge_hovered
-        {
-            self.schedule_menu_bar_close(cx);
-        }
-    }
-
-    pub(crate) fn set_menu_panel_hovered(&mut self, hovered: bool, cx: &mut Context<Self>) {
-        self.menu_panel_hovered = hovered;
-        if hovered {
-            self.menu_close_task = None;
-        } else if !self.menu_bar_hovered
-            && !self.menu_submenu_panel_hovered
-            && !self.menu_submenu_bridge_hovered
-        {
-            self.schedule_menu_bar_close(cx);
-        }
-    }
-
-    pub(crate) fn set_menu_submenu_panel_hovered(&mut self, hovered: bool, cx: &mut Context<Self>) {
-        self.menu_submenu_panel_hovered = hovered;
-        if hovered {
-            self.menu_close_task = None;
-        } else if !self.menu_bar_hovered
-            && !self.menu_panel_hovered
-            && !self.menu_submenu_bridge_hovered
-        {
-            self.schedule_menu_bar_close(cx);
-        }
-    }
-
-    /// Hover handler for the invisible gap bridge. The bridge and the submenu
-    /// panel overlap, so the cursor crossing between them fires a `false` for
-    /// one region and a `true` for the other in the same gesture. Keeping their
-    /// hover state in separate flags lets either one hold the menu open
-    /// regardless of the order those events arrive.
-    pub(crate) fn set_menu_submenu_bridge_hovered(
-        &mut self,
-        hovered: bool,
-        cx: &mut Context<Self>,
-    ) {
-        self.menu_submenu_bridge_hovered = hovered;
-        if hovered {
-            self.menu_close_task = None;
-        } else if !self.menu_bar_hovered
-            && !self.menu_panel_hovered
-            && !self.menu_submenu_panel_hovered
-        {
-            self.schedule_menu_bar_close(cx);
-        }
-    }
-
     pub(crate) fn dismiss_menu_bar_from_body(&mut self, cx: &mut Context<Self>) {
         if self.menu_bar_open.is_some() {
             self.close_menu_bar(cx);
         }
     }
-
-    // pub(crate) fn request_save_document(&mut self, cx: &mut Context<Self>) {
-    //     if !self.pending_save {
-    //         self.pending_save = true;
-    //         cx.notify();
-    //     }
-    // }
-
-    // pub(crate) fn request_save_document_as(&mut self, cx: &mut Context<Self>) {
-    //     if !self.pending_save_as {
-    //         self.pending_save_as = true;
-    //         cx.notify();
-    //     }
-    // }
 
     pub(crate) fn request_open_link_prompt(
         &mut self,
