+++ crates/opennote-velotype/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod app_menu;
 pub mod components;
 pub mod config;
 pub mod editor;
@@ -6,6 +5,4 @@ pub mod i18n;
 pub mod net;
 pub mod theme;
 
-mod app_identity;
 mod export;
-mod window_chrome;
